@@ -12,7 +12,7 @@ cmake_minimum_required(VERSION 3.10.0)
 # Language standard
 # NB: Seems like a good idea to have this before project() call.
 #
-set(CMAKE_CXX_STANDARD 14)
+set(CMAKE_CXX_STANDARD 17)
 set(CMAKE_CXX_STANDARD_REQUIRED ON)
 
 # ############################################################################ #
