@@ -1,23 +1,20 @@
-/* -*-  mode: c++; c-default-style: "google"; indent-tabs-mode: nil -*- */
-/**
- * @file   GenerationSpec.hh
- * @author William A. Perkins
- * @date Wed Sep 28 08:25:52 2011
- * 
- * @brief  Declaration of the GenerationSpec class
- * 
- * 
- */
+/*
+  Copyright 2010-201x held jointly by LANS/LANL, LBNL, and PNNL.
+  Amanzi is released under the three-clause BSD License.
+  The terms of use and "as is" disclaimer for this license are
+  provided in the top-level COPYRIGHT file.
 
-#ifndef __H_GenerationSpec
-#define __H_GenerationSpec
+  Authors: Rao Garimella, others
+*/
+
+#pragma once
 
 #include "Region.hh"
 #include "RegionBox.hh"
 #include "MeshDefs.hh"
 
 namespace Teuchos {
-class ParameterList;
+  class ParameterList;
 }
 
 namespace Amanzi {
@@ -31,11 +28,12 @@ namespace AmanziMesh {
  * The sole purpose of this class is to encapsulate the input
  * specifications for a generated unstructured mesh.  This should be
  * the only place in Amanzi where the parameter list for the
- * generation of an unstructured mesh is parsed.  
- * 
+ * generation of an unstructured mesh is parsed.
+ *
  */
 class GenerationSpec {
  public:
+
   /// Constructor that uses a parameter list
   GenerationSpec(const Teuchos::ParameterList& parameter_list);
 
@@ -43,7 +41,8 @@ class GenerationSpec {
   ~GenerationSpec(void);
 
   /// Get the overall spatial domain
-  const AmanziGeometry::RegionBox& domain(void) const { return *domain_; }
+  const AmanziGeometry::RegionBox& domain(void) const
+  { return *domain_; }
 
   /// Get the mesh dimensions in the x-direction
   const unsigned int& xcells(void) const { return nx_; }
@@ -55,37 +54,42 @@ class GenerationSpec {
   const unsigned int& zcells(void) const { return nz_; }
 
   /// Get access to the list of blocks or zones
-  AmanziGeometry::RegionVector::const_iterator block_begin(void) const { return blocks_.begin(); }
+  AmanziGeometry::RegionVector::const_iterator block_begin(void) const
+  { return blocks_.begin(); }
 
   /// Get access to the list of blocks or zones
-  AmanziGeometry::RegionVector::const_iterator block_end(void) const { return blocks_.end(); }
+  AmanziGeometry::RegionVector::const_iterator block_end(void) const
+  { return blocks_.end(); }
 
   Partitioner_type partitioner() const { return partitioner_; }
 
  protected:
+
   /// overall mesh domain  FIXME: We already have a domain
 
   Teuchos::RCP<AmanziGeometry::RegionBox> domain_;
 
-  unsigned int nx_; /**< number of cells in the x-direction */
-  unsigned int ny_; /**< number of cells in the y-direction */
-  unsigned int nz_; /**< number of cells in the y-direction */
+  unsigned int nx_;                     /**< number of cells in the x-direction */
+  unsigned int ny_;                     /**< number of cells in the y-direction */
+  unsigned int nz_;                     /**< number of cells in the y-direction */
 
   AmanziGeometry::RegionVector blocks_; /**< list of mesh subdomains */
 
-  Partitioner_type partitioner_ = PARTITIONER_DEFAULT; /**< partitioner type */
+  Partitioner_type partitioner_;
 
   /// fill attributes from specified list
-  void parse_(const Teuchos::ParameterList& parameter_list);
+  void parse_(const Teuchos::ParameterList &parameter_list);
 
  private:
+
   /// private, undefined copy constructor to avoid unwanted copies
 
   GenerationSpec(const GenerationSpec& old);
+
 };
 
 
 } // end namespace AmanziMesh
 } // end namespace Amanzi
 
-#endif
+
