@@ -1,13 +1,20 @@
-//
-// Mesh
-//
-// Copyright 2010-201x held jointly by LANL, ORNL, LBNL, and PNNL.
-// Amanzi is released under the three-clause BSD License.
-// The terms of use and "as is" disclaimer for this license are
-// provided in the top-level COPYRIGHT file.
-//
-// Helper functions for Mesh operations and algorithms
-//
+/*
+  Copyright 2010-201x held jointly by LANL, ORNL, LBNL, and PNNL.
+  Amanzi is released under the three-clause BSD License.
+  The terms of use and "as is" disclaimer for this license are
+  provided in the top-level COPYRIGHT file.
+
+  Authors: Ethan Coon (coonet@ornl.gov)
+*/
+// Functions for commonly used Mesh operations and algorithms
+/*
+
+  Note that these are distinct from Mesh_Helpers.  Helpers are functions most
+  commonly used within the mesh library, while these are functions that are
+  commonly used by clients of the mesh library.
+
+*/
+
 #pragma once
 
 #include "Epetra_MultiVector.h"
@@ -19,32 +26,32 @@ namespace AmanziMesh {
 // -----------------------------------------------------------------------------
 // Given a boundary face ID, get the corresponding face ID
 // -----------------------------------------------------------------------------
-AmanziMesh::Entity_ID
-getBoundaryFaceFace(const AmanziMesh::Mesh& mesh, AmanziMesh::Entity_ID bf);
+Entity_ID
+getBoundaryFaceFace(const Mesh& mesh, Entity_ID bf);
 
 // -----------------------------------------------------------------------------
 // Given a face ID, get the corresponding boundary face ID (assuming it is a bf)
 // -----------------------------------------------------------------------------
-AmanziMesh::Entity_ID
-getFaceOnBoundaryBoundaryFace(const AmanziMesh::Mesh& mesh, AmanziMesh::Entity_ID f);
+Entity_ID
+getFaceOnBoundaryBoundaryFace(const Mesh& mesh, Entity_ID f);
 
 // -----------------------------------------------------------------------------
 // Given a boundary face ID, get the cell internal to that face.
 // -----------------------------------------------------------------------------
-AmanziMesh::Entity_ID
-getBoundaryFaceInternalCell(const AmanziMesh::Mesh& mesh, AmanziMesh::Entity_ID bf);
+Entity_ID
+getBoundaryFaceInternalCell(const Mesh& mesh, Entity_ID bf);
 
 // -----------------------------------------------------------------------------
 // Given a face ID, and assuming it is a boundary face, get the cell internal.
 // -----------------------------------------------------------------------------
-AmanziMesh::Entity_ID
-getFaceOnBoundaryInternalCell(const AmanziMesh::Mesh& mesh, AmanziMesh::Entity_ID f);
+Entity_ID
+getFaceOnBoundaryInternalCell(const Mesh& mesh, Entity_ID f);
 
 // -----------------------------------------------------------------------------
 // Given a vector on faces, import to vector on boundary faces
 // -----------------------------------------------------------------------------
 void
-copyFacesToBoundaryFaces(const AmanziMesh::Mesh& mesh,
+copyFacesToBoundaryFaces(const Mesh& mesh,
                          const Epetra_MultiVector& faces,
                          Epetra_MultiVector& boundary_faces);
 
@@ -52,7 +59,7 @@ copyFacesToBoundaryFaces(const AmanziMesh::Mesh& mesh,
 // Given a vector on faces, import to vector on boundary faces
 // -----------------------------------------------------------------------------
 void
-copyBoundaryFacesToFaces(const AmanziMesh::Mesh& mesh,
+copyBoundaryFacesToFaces(const Mesh& mesh,
                          const Epetra_MultiVector& boundary_faces,
                          Epetra_MultiVector& faces);
 
@@ -60,24 +67,11 @@ copyBoundaryFacesToFaces(const AmanziMesh::Mesh& mesh,
 // Given a vector on cells, set the boundary_face entries by their internal cell
 // -----------------------------------------------------------------------------
 void
-copyCellsToBoundaryFaces(const AmanziMesh::Mesh& mesh,
+copyCellsToBoundaryFaces(const Mesh& mesh,
                          const Epetra_MultiVector& cells,
                          Epetra_MultiVector& boundary_faces);
 
 
-// -----------------------------------------------------------------------------
-// Given a boundary face f, return the exterior normal. If f is an interior face,
-// dir = 0 and normal orientation is not be reliable in parallel algorithms
-// -----------------------------------------------------------------------------
-AmanziGeometry::Point
-getFaceNormalExterior(const AmanziMesh::Mesh& mesh, int f, int* dir);
-
-
-// -----------------------------------------------------------------------------
-// Given a cell c and face f, returns the neighbooring cell
-// -----------------------------------------------------------------------------
-int
-cell_get_face_adj_cell(const AmanziMesh::Mesh& mesh, int c, int f);
 
 } // namespace AmanziMesh
 } // namespace Amanzi
