@@ -18,7 +18,8 @@ namespace AmanziMesh {
 const int MASTER = 1;
 const int GHOST = 2;
 
-} // namespace AmanziMesh
-} // namespace Amanzi
+}  // namespace AmanziMesh
+}  // namespace Amanzi
 
 #endif
+
