@@ -35,31 +35,31 @@ namespace AmanziMesh {
 * Light-weigthed constructor
 ****************************************************************** */
 MeshExtractedManifold::MeshExtractedManifold(
-  const Teuchos::RCP<const Mesh>& parent_mesh,
-  const std::string& setname,
-  const Entity_kind entity_kind,
-  const Comm_ptr_type& comm,
-  const Teuchos::RCP<const AmanziGeometry::GeometricModel>& gm,
-  const Teuchos::RCP<const Teuchos::ParameterList>& plist,
-  bool request_faces,
-  bool request_edges,
-  bool flattened)
-  : Mesh(comm, gm, plist, request_faces, request_edges),
+    const Teuchos::RCP<Mesh>& parent_mesh,
+    const std::string& setname, 
+    const Entity_kind entity_kind,
+    const Comm_ptr_type& comm,
+    const Teuchos::RCP<const AmanziGeometry::GeometricModel>& gm,
+    const Teuchos::RCP<Teuchos::ParameterList>& plist,
+    bool request_faces, bool request_edges,
+    bool flattened)
+  : MeshFramework(comm, gm, plist),
     parent_mesh_(parent_mesh),
     flattened_(flattened)
 {
+  
   vo_ = Teuchos::rcp(new VerboseObject(comm_, "MeshExtractedManifold", *plist_));
 
-  int d = parent_mesh_->space_dimension();
-  set_space_dimension(d);
-  set_manifold_dimension(d - 1);
-  if (flattened_) set_space_dimension(d - 1);
+  int d = parent_mesh_->getSpaceDimension();
+  setSpaceDimension(d);
+  setManifoldDimension(d - 1);
+  if (flattened_) setSpaceDimension(d - 1);
 
-  InitParentMaps(setname);
-  InitEpetraMaps();
+  InitParentMaps(setname); 
+  InitEpetraMaps(); 
   try {
     InitExteriorEpetraMaps();
-  } catch (...) {
+  } catch(...) {
     if (vo_.get() && vo_->os_OK(Teuchos::VERB_HIGH)) {
       Teuchos::OSTab tab = vo_->getOSTab();
       *(vo_->os()) << "Parent framework does not support buildng of exterior maps.\n";
@@ -73,11 +73,13 @@ MeshExtractedManifold::MeshExtractedManifold(
 /* ******************************************************************
 * Get cell type
 ****************************************************************** */
-Cell_type
-MeshExtractedManifold::cell_get_type(const Entity_ID c) const
+Cell_type MeshExtractedManifold::getCellType(const Entity_ID c) const
 {
-  int nfaces = cell_face_ids_[c].size();
-  return (nfaces == 4) ? QUAD : ((nfaces == 3) ? TRI : POLYGON);
+  Entity_ID_List faces; 
+  getCellFaces(c,faces);
+  int nfaces = faces.size();
+  return (nfaces == 4) ? Amanzi::AmanziMesh::Cell_type::QUAD : 
+    ((nfaces == 3) ? Amanzi::AmanziMesh::Cell_type::TRI : Amanzi::AmanziMesh::Cell_type::POLYGON);
 }
 
 
@@ -87,70 +89,31 @@ MeshExtractedManifold::cell_get_type(const Entity_ID c) const
 * Number of entities of any kind (cell, face, edge, node) and in a
 * particular category (OWNED, GHOST, ALL)
 ****************************************************************** */
-unsigned int
-MeshExtractedManifold::num_entities(const Entity_kind kind, const Parallel_type ptype) const
+std::size_t MeshExtractedManifold::getNumEntities(const Entity_kind kind, 
+                                                 const Parallel_type ptype) const
 {
   if (ptype == Parallel_type::OWNED)
     return nents_owned_[kind];
   else if (ptype == Parallel_type::ALL)
     return nents_owned_[kind] + nents_ghost_[kind];
 
-  return nents_ghost_[kind];
+  return nents_ghost_[kind]; 
 }
-
-
-/* ******************************************************************
-* Connectivity list: cell -> cells
-****************************************************************** */
-void
-MeshExtractedManifold::cell_get_face_adj_cells(const Entity_ID c,
-                                               const Parallel_type ptype,
-                                               Entity_ID_List* cells) const
-{
-  Entity_ID_List nodes, faces0, faces1;
-
-  int fp = entid_to_parent_[CELL][c];
-  parent_mesh_->face_get_nodes(fp, &nodes);
-  int nnodes = nodes.size();
-
-  int nfaces_p = parent_mesh_->num_entities(FACE, ptype);
-
-  cells->clear();
-  for (int i = 0; i < nnodes; ++i) {
-    int j = (i + 1) % nnodes;
-    int np0 = nodes[i];
-    int np1 = nodes[j];
-
-    parent_mesh_->node_get_faces(np0, ptype, &faces0);
-    parent_mesh_->node_get_faces(np1, ptype, &faces1);
-
-    for (auto it0 = faces0.begin(); it0 != faces0.end(); ++it0) {
-      for (auto it1 = faces1.begin(); it1 != faces1.end(); ++it1) {
-        if (*it0 == *it1 && *it0 != fp) {
-          if (*it0 < nfaces_p) cells->push_back(parent_to_entid_[CELL][*it0]);
-        }
-      }
-    }
-  }
-}
-
 
 /* ******************************************************************
 * Connectivity list: cell -> faces
 ****************************************************************** */
-void
-MeshExtractedManifold::cell_get_faces_and_dirs_internal_(const Entity_ID c,
-                                                         Entity_ID_List* faces,
-                                                         std::vector<int>* fdirs,
-                                                         const bool ordered) const
+void MeshExtractedManifold::getCellFacesAndDirs(
+    const Entity_ID c,
+    Entity_ID_List& faces, std::vector<int> *fdirs) const
 {
-  int fp = entid_to_parent_[CELL][c];
-  parent_mesh_->face_get_edges_and_dirs(fp, faces, fdirs);
-  int nfaces = faces->size();
-
+  int fp = entid_to_parent_[Entity_kind::CELL][c];
+  parent_mesh_->getFaceEdgesAndDirs(fp, faces, fdirs);
+  int nfaces = faces.size();
+ 
   for (int i = 0; i < nfaces; ++i) {
-    int f = (*faces)[i];
-    (*faces)[i] = parent_to_entid_[FACE][f];
+    int f = faces[i];
+    faces[i] = parent_to_entid_[Entity_kind::FACE][f];
   }
 
   // algorithms on a non-manifold use multiple normals and special continuity
@@ -158,8 +121,10 @@ MeshExtractedManifold::cell_get_faces_and_dirs_internal_(const Entity_ID c,
   // Now if the mesh is flattened, the Mesh class algorithm uses 2D edge
   // orientation. In this case, the result is correct iff the 3D face normal
   // is exterior.
-  if (!flattened_) {
-    for (int i = 0; i < nfaces; ++i) { (*fdirs)[i] = 1; }
+  if (! flattened_) {
+    for (int i = 0; i < nfaces; ++i) {
+      (*fdirs)[i] = 1;
+    }
   }
 }
 
@@ -167,18 +132,18 @@ MeshExtractedManifold::cell_get_faces_and_dirs_internal_(const Entity_ID c,
 /* ******************************************************************
 * Connectivity list: cell -> edges = cell -> faces
 ****************************************************************** */
-void
-MeshExtractedManifold::cell_get_edges_internal_(const Entity_ID c, Entity_ID_List* edges) const
+void MeshExtractedManifold::getCellEdges(
+    const Entity_ID c, Entity_ID_List& edges) const
 {
   std::vector<int> edirs;
 
-  int fp = entid_to_parent_[CELL][c];
-  parent_mesh_->face_get_edges_and_dirs(fp, edges, &edirs);
-  int nedges = edges->size();
+  int fp = entid_to_parent_[Entity_kind::CELL][c];
+  parent_mesh_->getFaceEdgesAndDirs(fp, edges, &edirs);
+  int nedges = edges.size();
 
   for (int i = 0; i < nedges; ++i) {
-    int e = (*edges)[i];
-    (*edges)[i] = parent_to_entid_[FACE][e];
+    int e = edges[i];
+    edges[i] = parent_to_entid_[Entity_kind::FACE][e];
   }
 }
 
@@ -186,16 +151,15 @@ MeshExtractedManifold::cell_get_edges_internal_(const Entity_ID c, Entity_ID_Lis
 /* ******************************************************************
 * Connectivity list: cell -> nodes
 ****************************************************************** */
-void
-MeshExtractedManifold::cell_get_nodes(const Entity_ID c, Entity_ID_List* nodes) const
+void MeshExtractedManifold::getCellNodes(const Entity_ID c, Entity_ID_List& nodes) const
 {
-  int fp = entid_to_parent_[CELL][c];
-  parent_mesh_->face_get_nodes(fp, nodes);
-  int nnodes = nodes->size();
+  int fp = entid_to_parent_[Entity_kind::CELL][c];
+  parent_mesh_->getFaceNodes(fp, nodes);
+  int nnodes = nodes.size();
 
   for (int i = 0; i < nnodes; ++i) {
-    int v = (*nodes)[i];
-    (*nodes)[i] = parent_to_entid_[NODE][v];
+    int v = nodes[i];
+    nodes[i] = parent_to_entid_[Entity_kind::NODE][v];
   }
 }
 
@@ -203,37 +167,33 @@ MeshExtractedManifold::cell_get_nodes(const Entity_ID c, Entity_ID_List* nodes) 
 /* ******************************************************************
 * Connectivity list: face -> nodes
 ****************************************************************** */
-void
-MeshExtractedManifold::face_get_nodes(const Entity_ID f, Entity_ID_List* nodes) const
-{
-  Entity_ID v0, v1;
+void MeshExtractedManifold::getFaceNodes(const Entity_ID f, Entity_ID_List& nodes) const {
+  Entity_ID_List v0;
 
-  int ep = entid_to_parent_[FACE][f];
-  parent_mesh_->edge_get_nodes(ep, &v0, &v1);
+  int ep = entid_to_parent_[Entity_kind::FACE][f];
+  v0 = parent_mesh_->getEdgeNodes(ep);
 
-  nodes->clear();
-  nodes->push_back(parent_to_entid_[NODE][v0]);
-  nodes->push_back(parent_to_entid_[NODE][v1]);
+  nodes.clear();
+  nodes.push_back(parent_to_entid_[Entity_kind::NODE][v0[0]]);
+  nodes.push_back(parent_to_entid_[Entity_kind::NODE][v0[1]]);
 }
 
 
 /* ******************************************************************
 * Connectivity list: face -> edges
 ****************************************************************** */
-void
-MeshExtractedManifold::face_get_edges_and_dirs_internal_(const Entity_ID f,
-                                                         Entity_ID_List* edges,
-                                                         std::vector<int>* edirs,
-                                                         const bool ordered) const
+void MeshExtractedManifold::getFaceEdgesAndDirs(
+    const Entity_ID f,
+    Entity_ID_List& edges, std::vector<int> *edirs) const
 {
-  Entity_ID v0, v1;
+  Entity_ID_List v0;
 
-  int ep = entid_to_parent_[FACE][f];
-  parent_mesh_->edge_get_nodes(ep, &v0, &v1);
+  int ep = entid_to_parent_[Entity_kind::FACE][f];
+  v0 = parent_mesh_->getEdgeNodes(ep);
 
-  edges->clear();
-  edges->push_back(parent_to_entid_[NODE][v0]);
-  edges->push_back(parent_to_entid_[NODE][v1]);
+  edges.clear();
+  edges.push_back(parent_to_entid_[Entity_kind::NODE][v0[0]]);
+  edges.push_back(parent_to_entid_[Entity_kind::NODE][v0[1]]);
   edirs->resize(2, 1);
 }
 
@@ -241,22 +201,21 @@ MeshExtractedManifold::face_get_edges_and_dirs_internal_(const Entity_ID f,
 /* ******************************************************************
 * Connectivity list: node -> cells
 ****************************************************************** */
-void
-MeshExtractedManifold::node_get_cells(const Entity_ID n,
-                                      const Parallel_type ptype,
-                                      Entity_ID_List* cells) const
+void MeshExtractedManifold::getNodeCells(
+    const Entity_ID n,
+    const Parallel_type ptype, Entity_ID_List& cells) const 
 {
   Entity_ID_List faces;
 
-  int np = entid_to_parent_[NODE][n];
-  parent_mesh_->node_get_faces(np, ptype, &faces);
+  int np = entid_to_parent_[Entity_kind::NODE][n];
+  parent_mesh_->getNodeFaces(np, ptype, faces);
   int nfaces = faces.size();
 
-  cells->clear();
+  cells.clear();
   for (int i = 0; i < nfaces; ++i) {
     int f = faces[i];
-    auto it = parent_to_entid_[CELL].find(f);
-    if (it != parent_to_entid_[CELL].end()) cells->push_back(it->second);
+    auto it = parent_to_entid_[Entity_kind::CELL].find(f);
+    if (it != parent_to_entid_[Entity_kind::CELL].end()) cells.push_back(it->second);
   }
 }
 
@@ -264,22 +223,20 @@ MeshExtractedManifold::node_get_cells(const Entity_ID n,
 /* ******************************************************************
 * Connectivity list: edge -> cells
 ****************************************************************** */
-void
-MeshExtractedManifold::edge_get_cells(const Entity_ID e,
-                                      const Parallel_type ptype,
-                                      Entity_ID_List* cells) const
+void MeshExtractedManifold::getEdgeCells(
+   const Entity_ID e, const Parallel_type ptype, Entity_ID_List& cells) const 
 {
   Entity_ID_List faces;
 
-  int ep = entid_to_parent_[FACE][e];
-  parent_mesh_->edge_get_faces(ep, ptype, &faces);
+  int ep = entid_to_parent_[Entity_kind::FACE][e];
+  parent_mesh_->getEdgeFaces(ep, ptype, faces);
   int nfaces = faces.size();
 
-  cells->clear();
+  cells.clear();
   for (int i = 0; i < nfaces; ++i) {
     int f = faces[i];
-    auto it = parent_to_entid_[CELL].find(f);
-    if (it != parent_to_entid_[CELL].end()) cells->push_back(it->second);
+    auto it = parent_to_entid_[Entity_kind::CELL].find(f);
+    if (it != parent_to_entid_[Entity_kind::CELL].end()) cells.push_back(it->second);
   }
 }
 
@@ -287,130 +244,192 @@ MeshExtractedManifold::edge_get_cells(const Entity_ID e,
 /* ******************************************************************
 * Connectivity list: face -> cells
 ****************************************************************** */
-void
-MeshExtractedManifold::face_get_cells_internal_(const Entity_ID f,
-                                                const Parallel_type ptype,
-                                                Entity_ID_List* cells) const
+void MeshExtractedManifold::getFaceCells(
+    const Entity_ID f,
+    const Parallel_type ptype, Entity_ID_List& cells) const
 {
   Entity_ID_List faces;
 
-  int ep = entid_to_parent_[FACE][f];
-  parent_mesh_->edge_get_faces(ep, ptype, &faces);
+  int ep = entid_to_parent_[Entity_kind::FACE][f];
+  parent_mesh_->getEdgeFaces(ep, ptype, faces);
   int nfaces = faces.size();
 
-  cells->clear();
+  cells.clear();
   for (int i = 0; i < nfaces; ++i) {
-    auto it = parent_to_entid_[CELL].find(faces[i]);
-    if (it != parent_to_entid_[CELL].end()) cells->push_back(it->second);
+    auto it = parent_to_entid_[Entity_kind::CELL].find(faces[i]);
+    if (it != parent_to_entid_[Entity_kind::CELL].end()) cells.push_back(it->second);
   }
+}
+
+
+/* ******************************************************************
+* The node coordinates are returned in in arbitrary order
+****************************************************************** */
+Point_List MeshExtractedManifold::getCellCoordinates(
+    const Entity_ID c) const
+{
+  Entity_ID_List nodes; 
+
+  Point_List vxyz; 
+  int fp = entid_to_parent_[Entity_kind::CELL][c];
+  parent_mesh_->getFaceNodes(fp, nodes);
+  int nnodes = nodes.size();
+
+  AmanziGeometry::Point p(getSpaceDimension());
+
+  vxyz.clear();
+  for (int i = 0; i < nnodes; ++i) {
+    p = parent_mesh_->getNodeCoordinate(nodes[i]);
+    vxyz.push_back(p);
+  }
+
+  if (flattened_) {
+    for (int i = 0; i < nnodes; ++i) {
+      vxyz[i].set(vxyz[i][0], vxyz[i][1]);
+    }
+  }
+  return vxyz; 
+}
+
+
+/* ******************************************************************
+* Face coordinates use convention same as in face_to_nodes()
+****************************************************************** */
+Point_List MeshExtractedManifold::getFaceCoordinates(
+    const Entity_ID f) const
+{
+  Point_List vxyz; 
+  Entity_ID_List v0;
+
+  int ep = entid_to_parent_[Entity_kind::FACE][f];
+  v0 = parent_mesh_->getEdgeNodes(ep);
+
+  AmanziGeometry::Point xyz(getSpaceDimension());
+
+  vxyz.clear();
+  xyz = parent_mesh_->getNodeCoordinate(v0[0]);
+  vxyz.push_back(xyz);
+
+  xyz = parent_mesh_->getNodeCoordinate(v0[1]);
+  vxyz.push_back(xyz);
+
+  if (flattened_) {
+    for (int i = 0; i < 2; ++i) {
+      vxyz[i].set(vxyz[i][0], vxyz[i][1]);
+    }
+  }
+  return vxyz; 
 }
 
 
 /* ******************************************************************
 * Position vector for a node
 ****************************************************************** */
-void
-MeshExtractedManifold::node_get_coordinates(const Entity_ID n, AmanziGeometry::Point* xyz) const
+AmanziGeometry::Point MeshExtractedManifold::getNodeCoordinate(
+    const Entity_ID n) const
 {
-  auto np = entid_to_parent_[NODE][n];
-  parent_mesh_->node_get_coordinates(np, xyz);
+  AmanziGeometry::Point xyz; 
+  auto np = entid_to_parent_[Entity_kind::NODE][n];
+  xyz = parent_mesh_->getNodeCoordinate(np);
 
-  if (flattened_) xyz->set((*xyz)[0], (*xyz)[1]);
+  if (flattened_) xyz.set(xyz[0], xyz[1]);
+  return xyz; 
 }
 
 
 /* ******************************************************************
 * Get list of entities of type 'ptype' in set specified by setname
 ****************************************************************** */
-void
-MeshExtractedManifold::get_set_entities_and_vofs(const std::string& setname,
-                                                 const Entity_kind kind,
-                                                 const Parallel_type ptype,
-                                                 std::vector<Entity_ID>* setents,
-                                                 std::vector<double>* vofs) const
+decltype(auto) MeshExtractedManifold::getSetEntitiesAndVolumeFractions(
+    const std::string& setname, 
+    const Entity_kind kind, const Parallel_type ptype ) const
 {
-  assert(setents != nullptr);
 
-  setents->clear();
-  Teuchos::RCP<const AmanziGeometry::GeometricModel> gm = geometric_model();
+  std::vector<Entity_ID> setents; 
+  std::vector<double> vofs; 
+  
+  Teuchos::RCP<const AmanziGeometry::GeometricModel> gm = getGeometricModel();
 
   // is there an appropriate region by this name?
   Teuchos::RCP<const AmanziGeometry::Region> rgn;
   try {
     rgn = gm->FindRegion(setname);
   } catch (...) {
-    valid_set_name(setname, kind);
+    // Catch
   }
 
   if (rgn == Teuchos::null) {
     std::stringstream ss;
-    ss << "Geometric model has no region named \"" << setname << "\", kind=" << kind << "\n";
+    ss << "Geometric model has no region named \"" << setname <<"\", kind=" << static_cast<int>(kind) << "\n";
     Errors::Message msg(ss.str());
     Exceptions::amanzi_throw(msg);
   }
 
-  std::string setname_internal = setname + std::to_string(kind);
-  if (sets_.find(setname_internal) != sets_.end()) { *setents = sets_[setname_internal]; }
+  std::string setname_internal = setname + std::to_string(static_cast<int>(kind));
+  if (sets_.find(setname_internal) != sets_.end()) {
+    setents = sets_[setname_internal];
+  }
 
   // set does not exist - build it
   if (sets_.find(setname_internal) == sets_.end()) {
     sets_[setname_internal] = build_set_(rgn, kind);
-    *setents = sets_[setname_internal];
+    setents = sets_[setname_internal];
   }
 
   // extract entities only of specified parallel type
   int m(0);
 
   if (ptype == Parallel_type::ALL) {
-    m = setents->size();
+    m = setents.size();
   } else if (ptype == Parallel_type::OWNED) {
-    int nents = num_entities(kind, ptype);
-    for (int n = 0; n < setents->size(); ++n) {
-      if ((*setents)[n] < nents) (*setents)[m++] = (*setents)[n];
+    int nents = getNumEntities(kind, ptype); 
+    for (int n = 0; n < setents.size(); ++n) {
+      if ((setents)[n] < nents) (setents)[m++] = (setents)[n];
     }
   } else if (ptype == Parallel_type::GHOST) {
-    int nents = num_entities(kind, Parallel_type::OWNED);
-    for (int n = 0; n < setents->size(); ++n) {
-      if ((*setents)[n] >= nents) (*setents)[m++] = (*setents)[n];
+    int nents = getNumEntities(kind, Parallel_type::OWNED); 
+    for (int n = 0; n < setents.size(); ++n) {
+      if ((setents)[n] >= nents) (setents)[m++] = (setents)[n];
     }
   }
-
-  setents->resize(m);
-
+    
+  setents.resize(m);
+      
   // Check if there were no entities left on any processor after
   // extracting the appropriate category of entities
-  int mglob;
-  get_comm()->SumAll(&m, &mglob, 1);
-
+  int mglob; 
+  getComm()->SumAll(&m, &mglob, 1);
+  
   if (mglob == 0) {
     std::stringstream ss;
-    ss << "Could not retrieve mesh entities of kind=" << kind << " for set \"" << setname << "\""
-       << std::endl;
+    ss << "Could not retrieve mesh entities of kind=" << static_cast<int>(kind) << " for set \"" << setname << "\"" << std::endl;
     Errors::Message msg(ss.str());
     Exceptions::amanzi_throw(msg);
   }
+
+  return std::make_pair(setents, vofs);
+
 }
 
 
 /* ******************************************************************
 * Create a set from similar set in mesh
 ****************************************************************** */
-Entity_ID_List
-MeshExtractedManifold::build_set_(const Teuchos::RCP<const AmanziGeometry::Region>& rgn,
-                                  const Entity_kind kind) const
+Entity_ID_List MeshExtractedManifold::build_set_(
+    const Teuchos::RCP<const AmanziGeometry::Region>& rgn, const Entity_kind kind) const
 {
   // modify rgn/set name by prefixing it with the type of entity requested
-  std::string internal_name = rgn->get_name() + std::to_string(kind);
+  std::string internal_name = rgn->get_name() + std::to_string(static_cast<int>(kind));
 
-  // create entity set based on the region definition
+  // create entity set based on the region definition  
   Entity_ID_List mset;
 
   // special processing of regions
   if (rgn->get_type() == AmanziGeometry::RegionType::LABELEDSET) {
     if (parent_labeledsets_.find(internal_name) == parent_labeledsets_.end()) {
-      Entity_kind kind_p(NODE);
-      if (kind == CELL) kind_p = FACE;
-      if (kind == FACE) kind_p = EDGE;
+      Entity_kind kind_p(Entity_kind::NODE);
+      if (kind == Entity_kind::CELL) kind_p = Entity_kind::FACE;
+      if (kind == Entity_kind::FACE) kind_p = Entity_kind::EDGE;
 
       Entity_ID_List setents;
       TryExtension_(rgn->get_name(), kind_p, kind, &setents);
@@ -424,9 +443,11 @@ MeshExtractedManifold::build_set_(const Teuchos::RCP<const AmanziGeometry::Regio
   }
 
   // generic algorithm
-  int nents_wghost = num_entities(kind, Parallel_type::ALL);
-  if (rgn->get_type() == AmanziGeometry::RegionType::ALL) {
-    for (int n = 0; n < nents_wghost; ++n) { mset.push_back(n); }
+  int nents_wghost = getNumEntities(kind, Parallel_type::ALL);
+  if (rgn->get_type() == AmanziGeometry::RegionType::ALL)  {
+    for (int n = 0; n < nents_wghost; ++n) {
+      mset.push_back(n);
+    }
     return mset;
   }
 
@@ -438,11 +459,11 @@ MeshExtractedManifold::build_set_(const Teuchos::RCP<const AmanziGeometry::Regio
     if (mset.size() > 0) return mset;
   }
 
-  if (kind == CELL)
+  if (kind == Entity_kind::CELL)
     mset = build_set_cells_(rgn, &missing);
-  else if (kind == FACE)
+  else if (kind == Entity_kind::FACE) 
     mset = build_set_faces_(rgn, &missing);
-  else if (kind == NODE)
+  else if (kind == Entity_kind::NODE) 
     mset = build_set_nodes_(rgn, &missing);
 
   // check if this set exists in parent mesh
@@ -454,21 +475,21 @@ MeshExtractedManifold::build_set_(const Teuchos::RCP<const AmanziGeometry::Regio
   if (missing) {
     if (vo_.get() && vo_->os_OK(Teuchos::VERB_HIGH)) {
       Teuchos::OSTab tab = vo_->getOSTab();
-      *(vo_->os()) << "Requested entities of kind=" << kind << " on region=\"" << rgn->get_name()
-                   << "\" of type " << rgn->get_type() << " and dimension "
-                   << rgn->get_manifold_dimension() << ". Result is an empty set.\n";
+      *(vo_->os()) << "Requested entities of kind=" << static_cast<int>(kind) 
+        << " on region=\"" << rgn->get_name() << "\" of type " << rgn->get_type()  
+        << " and dimension " << rgn->get_manifold_dimension() << ". Result is an empty set.\n";
     }
   }
 
   if (rgn->get_type() == AmanziGeometry::RegionType::LOGICAL) {
-    Teuchos::RCP<const AmanziGeometry::GeometricModel> gm = geometric_model();
+    Teuchos::RCP<const AmanziGeometry::GeometricModel> gm = getGeometricModel();
 
     auto boolrgn = Teuchos::rcp_static_cast<const AmanziGeometry::RegionLogical>(rgn);
     const std::vector<std::string> rgn_names = boolrgn->get_component_regions();
     int nregs = rgn_names.size();
-
-    std::vector<std::set<Entity_ID>> msets;
-
+    
+    std::vector<std::set<Entity_ID> > msets;
+    
     for (int r = 0; r < nregs; r++) {
       auto rgn1 = gm->FindRegion(rgn_names[r]);
 
@@ -479,49 +500,54 @@ MeshExtractedManifold::build_set_(const Teuchos::RCP<const AmanziGeometry::Regio
         Errors::Message msg(ss.str());
         Exceptions::amanzi_throw(msg);
       }
-
-      std::string setname_internal = rgn1->get_name() + std::to_string(kind);
+        
+      std::string setname_internal = rgn1->get_name() + std::to_string(static_cast<int>(kind));
 
       if (sets_.find(setname_internal) == sets_.end())
-        sets_[setname_internal] = build_set_(rgn1, kind);
-
+        sets_[setname_internal] = build_set_(rgn1, kind); 
+     
       auto it = sets_.find(setname_internal);
-      msets.push_back(std::set<Entity_ID>(it->second.begin(), it->second.end()));
+      msets.push_back(std::set<Entity_ID>(it->second.begin(), it->second.end())); 
     }
 
     if (boolrgn->get_operation() == AmanziGeometry::BoolOpType::UNION) {
       for (int n = 1; n < msets.size(); ++n) {
-        for (auto it = msets[n].begin(); it != msets[n].end(); ++it) msets[0].insert(*it);
+        for (auto it = msets[n].begin(); it != msets[n].end(); ++it)
+          msets[0].insert(*it);
       }
     }
 
     else if (boolrgn->get_operation() == AmanziGeometry::BoolOpType::SUBTRACT) {
       for (int n = 2; n < msets.size(); ++n) {
-        for (auto it = msets[n].begin(); it != msets[n].end(); ++it) msets[0].insert(*it);
+        for (auto it = msets[n].begin(); it != msets[n].end(); ++it)
+          msets[0].insert(*it);
       }
 
-      for (auto it = msets[1].begin(); it != msets[1].end(); ++it) msets[0].erase(*it);
+      for (auto it = msets[1].begin(); it != msets[1].end(); ++it)
+        msets[0].erase(*it);
     }
 
     else if (boolrgn->get_operation() == AmanziGeometry::BoolOpType::COMPLEMENT) {
       for (int n = 1; n < msets.size(); ++n) {
-        for (auto it = msets[n].begin(); it != msets[n].end(); ++it) msets[0].insert(*it);
-      }
+        for (auto it = msets[n].begin(); it != msets[n].end(); ++it)
+          msets[0].insert(*it);
+      } 
 
       auto tmp = msets[0];
       msets[0].clear();
       for (int n = 0; n < nents_wghost; ++n)
         if (tmp.find(n) == tmp.end()) msets[0].insert(n);
-    }
+    } 
 
     else if (boolrgn->get_operation() == AmanziGeometry::BoolOpType::INTERSECT) {
       for (int n = 1; n < msets.size(); ++n) {
         for (auto it = msets[n].begin(); it != msets[n].end(); ++it)
           if (msets[0].find(*it) == msets[0].end()) msets[0].erase(*it);
       }
-    }
+    } 
 
-    for (auto it = msets[0].begin(); it != msets[0].end(); ++it) mset.push_back(*it);
+    for (auto it = msets[0].begin(); it != msets[0].end(); ++it)
+      mset.push_back(*it);
   }
 
   return mset;
@@ -531,45 +557,43 @@ MeshExtractedManifold::build_set_(const Teuchos::RCP<const AmanziGeometry::Regio
 /* ******************************************************************
 * Create a cell-set from similar set in mesh
 ****************************************************************** */
-Entity_ID_List
-MeshExtractedManifold::build_set_cells_(const Teuchos::RCP<const AmanziGeometry::Region>& rgn,
-                                        bool* missing) const
+Entity_ID_List MeshExtractedManifold::build_set_cells_(
+    const Teuchos::RCP<const AmanziGeometry::Region>& rgn, bool* missing) const
 {
-  int space_dim = space_dimension();
-  int manifold_dim = manifold_dimension();
+  int space_dim = getSpaceDimension();
+  int manifold_dim = getManifoldDimension();
 
   Entity_ID_List mset;
 
-  int ncells_wghost = num_entities(CELL, Parallel_type::ALL);
+  int ncells_wghost = getNumEntities(Entity_kind::CELL, Parallel_type::ALL);              
   *missing = false;
 
   if (rgn->get_type() == AmanziGeometry::RegionType::BOX ||
       rgn->get_type() == AmanziGeometry::RegionType::CYLINDER ||
       rgn->get_type() == AmanziGeometry::RegionType::COLORFUNCTION) {
+
     for (int c = 0; c < ncells_wghost; ++c)
-      if (rgn->inside(cell_centroid(c))) mset.push_back(c);
+      if (rgn->inside(getCellCentroid(c))) mset.push_back(c);
   }
 
   else if (rgn->get_type() == AmanziGeometry::RegionType::POINT) {
     auto rp = Teuchos::rcp_static_cast<const AmanziGeometry::RegionPoint>(rgn)->point();
 
     for (int c = 0; c < ncells_wghost; ++c)
-      if (PointInCell_(rp, c)) mset.push_back(c);
+      assert(false); 
+      //if (point_in_cell(rp, c)) mset.push_back(c);
   }
 
   else if (((rgn->get_type() == AmanziGeometry::RegionType::PLANE) ||
-            (rgn->get_type() == AmanziGeometry::RegionType::POLYGON)) &&
+            (rgn->get_type() == AmanziGeometry::RegionType::POLYGON)) && 
            manifold_dim == 2) {
-    Entity_ID_List nodes;
-    AmanziGeometry::Point xyz(space_dim);
-
     for (int c = 0; c < ncells_wghost; ++c) {
-      cell_get_nodes(c, &nodes);
+      std::vector<AmanziGeometry::Point> ccoords(space_dim);
+      ccoords = getCellCoordinates(c);
 
       bool on_plane(true);
-      for (int i = 0; i < nodes.size(); ++i) {
-        node_get_coordinates(nodes[i], &xyz);
-        if (!rgn->inside(xyz)) {
+      for (int i = 0; i < ccoords.size(); ++i) {
+        if (!rgn->inside(ccoords[i])) {
           on_plane = false;
           break;
         }
@@ -589,12 +613,11 @@ MeshExtractedManifold::build_set_cells_(const Teuchos::RCP<const AmanziGeometry:
 /* ******************************************************************
 * Create a face-set from similar set in mesh
 ****************************************************************** */
-Entity_ID_List
-MeshExtractedManifold::build_set_faces_(const Teuchos::RCP<const AmanziGeometry::Region>& rgn,
-                                        bool* missing) const
+Entity_ID_List MeshExtractedManifold::build_set_faces_(
+    const Teuchos::RCP<const AmanziGeometry::Region>& rgn, bool* missing) const
 {
-  int space_dim = space_dimension();
-  int nfaces_wghost = num_entities(FACE, Parallel_type::ALL);
+  int space_dim = getSpaceDimension();
+  int nfaces_wghost = getNumEntities(Entity_kind::FACE, Parallel_type::ALL);              
 
   Entity_ID_List mset;
 
@@ -603,22 +626,19 @@ MeshExtractedManifold::build_set_faces_(const Teuchos::RCP<const AmanziGeometry:
   if (rgn->get_type() == AmanziGeometry::RegionType::BOX ||
       rgn->get_type() == AmanziGeometry::RegionType::CYLINDER) {
     for (int f = 0; f < nfaces_wghost; ++f) {
-      if (rgn->inside(face_centroid(f))) mset.push_back(f);
+      if (rgn->inside(getFaceCentroid(f))) mset.push_back(f);
     }
   }
 
   else if (rgn->get_type() == AmanziGeometry::RegionType::PLANE ||
            rgn->get_type() == AmanziGeometry::RegionType::POLYGON) {
     for (int f = 0; f < nfaces_wghost; ++f) {
-      Entity_ID_List nodes;
-      face_get_nodes(f, &nodes);
-
-      AmanziGeometry::Point xp(space_dim);
-
+      std::vector<AmanziGeometry::Point> fcoords(space_dim);
+      fcoords = getFaceCoordinates(f);
+            
       bool on_plane(true);
-      for (int i = 0; i < nodes.size(); ++i) {
-        node_get_coordinates(nodes[i], &xp);
-        if (!rgn->inside(xp)) {
+      for (int i = 0; i < fcoords.size(); ++i) {
+        if (!rgn->inside(fcoords[i])) {
           on_plane = false;
           break;
         }
@@ -627,14 +647,14 @@ MeshExtractedManifold::build_set_faces_(const Teuchos::RCP<const AmanziGeometry:
     }
   }
 
-  else if (rgn->get_type() == AmanziGeometry::RegionType::BOUNDARY) {
-    const Epetra_Map& fmap = face_map(true);
-    const Epetra_Map& map = exterior_face_map(true);
+  else if (rgn->get_type() == AmanziGeometry::RegionType::BOUNDARY)  {
+    const auto& fmap = ent_map_owned_[Entity_kind::FACE]; 
+    const auto& map = ent_extmap_owned_[Entity_kind::FACE]; 
 
-    int nfaces = map.NumMyElements();
+    int nfaces = map->NumMyElements(); 
 
     for (int f = 0; f < nfaces; ++f) {
-      int lid = fmap.LID(map.GID(f));
+      int lid = fmap->LID(map->GID(f));
       mset.push_back(lid);
     }
   }
@@ -650,12 +670,11 @@ MeshExtractedManifold::build_set_faces_(const Teuchos::RCP<const AmanziGeometry:
 /* ******************************************************************
 * Create a node-set from similar set in mesh
 ****************************************************************** */
-Entity_ID_List
-MeshExtractedManifold::build_set_nodes_(const Teuchos::RCP<const AmanziGeometry::Region>& rgn,
-                                        bool* missing) const
+Entity_ID_List MeshExtractedManifold::build_set_nodes_(
+    const Teuchos::RCP<const AmanziGeometry::Region>& rgn, bool* missing) const
 {
-  int space_dim = space_dimension();
-  int nnodes_wghost = num_entities(NODE, Parallel_type::ALL);
+  int space_dim = getSpaceDimension();
+  int nnodes_wghost = getNumEntities(Entity_kind::NODE, Parallel_type::ALL);
 
   Entity_ID_List mset;
 
@@ -668,24 +687,24 @@ MeshExtractedManifold::build_set_nodes_(const Teuchos::RCP<const AmanziGeometry:
       rgn->get_type() == AmanziGeometry::RegionType::POINT) {
     for (int v = 0; v < nnodes_wghost; ++v) {
       AmanziGeometry::Point xp(space_dim);
-      node_get_coordinates(v, &xp);
-
+      xp = getNodeCoordinate(v);
+                  
       if (rgn->inside(xp)) {
         mset.push_back(v);
         // Only one node per point rgn
-        if (rgn->get_type() == AmanziGeometry::RegionType::POINT) break;
+        if (rgn->get_type() == AmanziGeometry::RegionType::POINT) break;      
       }
     }
   }
 
-  else if (rgn->get_type() == AmanziGeometry::RegionType::BOUNDARY) {
-    const Epetra_Map& vmap = node_map(true);
-    const Epetra_Map& map = exterior_node_map(true);
+  else if (rgn->get_type() == AmanziGeometry::RegionType::BOUNDARY)  {
+    const auto& vmap = ent_map_owned_[Entity_kind::NODE]; 
+    const auto& map = ent_extmap_owned_[Entity_kind::NODE]; 
 
-    int nnodes = map.NumMyElements();
+    int nnodes = ent_map_owned_[Entity_kind::NODE]->NumMyElements(); 
 
     for (int v = 0; v < nnodes; ++v) {
-      int lid = vmap.LID(map.GID(v));
+      int lid = vmap->LID(map->GID(v));
       mset.push_back(lid);
     }
   }
@@ -701,24 +720,23 @@ MeshExtractedManifold::build_set_nodes_(const Teuchos::RCP<const AmanziGeometry:
 /* ******************************************************************
 * Create from a parent mesh
 ****************************************************************** */
-Entity_ID_List
-MeshExtractedManifold::build_from_parent_(const std::string& rgnname,
-                                          const Entity_kind kind_d) const
+Entity_ID_List MeshExtractedManifold::build_from_parent_(
+  const std::string& rgnname, const Entity_kind kind_d) const
 {
   Entity_ID_List mset, mset_tmp;
   Entity_kind kind_p;
 
-  if (kind_d == CELL)
-    kind_p = FACE;
-  else if (kind_d == FACE)
-    kind_p = EDGE;
-  else if (kind_d == NODE)
-    kind_p = NODE;
-  else
+  if (kind_d == Entity_kind::CELL) 
+    kind_p = Entity_kind::FACE;
+  else if (kind_d == Entity_kind::FACE) 
+    kind_p = Entity_kind::EDGE;
+  else if (kind_d == Entity_kind::NODE)
+    kind_p = Entity_kind::NODE;
+  else 
     return mset;
 
   try {
-    parent_mesh_->get_set_entities(rgnname, kind_p, Parallel_type::ALL, &mset_tmp);
+    mset_tmp = parent_mesh_->getSetEntities(rgnname, kind_p, Parallel_type::ALL);
 
     for (int i = 0; i < mset_tmp.size(); ++i) {
       int f = mset_tmp[i];
@@ -736,16 +754,15 @@ MeshExtractedManifold::build_from_parent_(const std::string& rgnname,
 /* ******************************************************************
 * Create internal maps for child->parent
 ****************************************************************** */
-void
-MeshExtractedManifold::InitParentMaps(const std::string& setname)
+void MeshExtractedManifold::InitParentMaps(const std::string& setname)
 {
   nents_owned_.clear();
   nents_ghost_.clear();
   entid_to_parent_.clear();
   parent_to_entid_.clear();
 
-  std::vector<Entity_kind> kinds_extracted({ CELL, FACE, NODE });
-  std::vector<Entity_kind> kinds_parent({ FACE, EDGE, NODE });
+  std::vector<Entity_kind> kinds_extracted({Entity_kind::CELL, Entity_kind::FACE, Entity_kind::NODE});
+  std::vector<Entity_kind> kinds_parent({Entity_kind::FACE, Entity_kind::EDGE, Entity_kind::NODE});
 
   for (int i = 0; i < 3; ++i) {
     auto kind_d = kinds_extracted[i];
@@ -757,7 +774,7 @@ MeshExtractedManifold::InitParentMaps(const std::string& setname)
 
     TryExtension_(setname, kind_p, kind_d, &setents);
     if (setents.size() == 0)
-      parent_mesh_->get_set_entities_and_vofs(setname, kind_p, Parallel_type::ALL, &setents, &vofs);
+      std::tie(setents,vofs) = getSetEntitiesAndVolumeFractions(setname, kind_p, Parallel_type::ALL);
 
     auto marked_ents = EnforceOneLayerOfGhosts_(setname, kind_p, &setents);
 
@@ -779,7 +796,9 @@ MeshExtractedManifold::InitParentMaps(const std::string& setname)
     // create reverse ordered map
     auto& ids_d = parent_to_entid_[kind_d];
     ids_d.clear();
-    for (int n = 0; n < ids_p.size(); ++n) { ids_d[ids_p[n]] = n; }
+    for (int n = 0; n < ids_p.size(); ++n) {
+      ids_d[ids_p[n]] = n;
+    }
   }
 }
 
@@ -789,21 +808,18 @@ MeshExtractedManifold::InitParentMaps(const std::string& setname)
 * owned or used by this processor. This helps Epetra understand 
 * inter-partition dependencies of the data.
 ****************************************************************** */
-void
-MeshExtractedManifold::InitEpetraMaps()
+void MeshExtractedManifold::InitEpetraMaps()
 {
-  std::vector<Entity_kind> kinds_extracted({ CELL, FACE, NODE });
-  std::vector<Entity_kind> kinds_parent({ FACE, EDGE, NODE });
+  std::vector<Entity_kind> kinds_extracted({Entity_kind::CELL, Entity_kind::FACE, Entity_kind::NODE});
+  std::vector<Entity_kind> kinds_parent({Entity_kind::FACE, Entity_kind::EDGE, Entity_kind::NODE});
 
   for (int i = 0; i < 3; ++i) {
     auto kind_d = kinds_extracted[i];
     auto kind_p = kinds_parent[i];
 
-    // compute (discontinuous) owned global ids using the parent map
-    Teuchos::RCP<const Epetra_BlockMap> parent_map =
-      Teuchos::rcpFromRef(parent_mesh_->map(kind_p, false));
-    Teuchos::RCP<const Epetra_BlockMap> parent_map_wghost =
-      Teuchos::rcpFromRef(parent_mesh_->map(kind_p, true));
+    // compute (discontinuous) owned global ids using the parent map 
+    auto parent_map = Teuchos::rcpFromRef(parent_to_entid_[kind_p]);
+    auto parent_map_wghost = Teuchos::rcpFromRef(parent_to_entid_[kind_p]);
 
     int nents = nents_owned_[kind_d];
     int nents_wghost = nents_owned_[kind_d] + nents_ghost_[kind_d];
@@ -811,7 +827,7 @@ MeshExtractedManifold::InitEpetraMaps()
 
     for (int n = 0; n < nents; ++n) {
       int id = entid_to_parent_[kind_d][n];
-      gids[n] = parent_map_wghost->GID(id);
+      gids[n] = parent_map_wghost->at(id);
     }
 
     auto subset_map = Teuchos::rcp(new Epetra_Map(-1, nents, gids, 0, *comm_));
@@ -819,20 +835,23 @@ MeshExtractedManifold::InitEpetraMaps()
     // compute owned + ghost ids using the parent map and the minimum global id
     for (int n = 0; n < nents_wghost; ++n) {
       int id = entid_to_parent_[kind_d][n];
-      gids[n] = parent_map_wghost->GID(id);
+      gids[n] = parent_map_wghost->at(id);
     }
 
     auto subset_map_wghost = Teuchos::rcp(new Epetra_Map(-1, nents_wghost, gids, 0, *comm_));
-    delete[] gids;
+    delete [] gids;
 
     // creare continuous maps
     auto mymesh = Teuchos::rcpFromRef(*this);
-    auto tmp = CreateContinuousMaps(mymesh,
-                                    std::make_pair(parent_map, parent_map_wghost),
-                                    std::make_pair(subset_map, subset_map_wghost));
+    Mesh mcache(mymesh,plist_); 
+    assert(false); 
+    // Which method to use here? 
+    //auto tmp = createContinuousMaps(Teuchos::rcpFromRef(mcache),
+    //                                std::make_pair(parent_map, parent_map_wghost),
+    //                                std::make_pair(subset_map, subset_map_wghost));
 
-    ent_map_owned_[kind_d] = tmp.first;
-    ent_map_wghost_[kind_d] = tmp.second;
+    //ent_map_owned_[kind_d] = tmp.first;
+    //ent_map_wghost_[kind_d] = tmp.second;
   }
 }
 
@@ -841,98 +860,94 @@ MeshExtractedManifold::InitEpetraMaps()
 * Exterior Epetra maps are cannot be always extracted from a
 * parent mesh, so we build them explicitly.
 ****************************************************************** */
-void
-MeshExtractedManifold::InitExteriorEpetraMaps()
+void MeshExtractedManifold::InitExteriorEpetraMaps()
 {
-  int nents = nents_owned_[FACE];
+  int nents = nents_owned_[Entity_kind::FACE];
   Entity_ID_List cells, nodes;
 
-  const auto& fmap_owned = face_map(false);
-  const auto& fmap_wghost = face_map(true);
+  const auto& fmap_owned = ent_map_owned_[Entity_kind::FACE];
+  const auto& fmap_wghost = ent_map_wghost_[Entity_kind::FACE];
 
-  const auto& vmap_owned = node_map(false);
-  const auto& vmap_wghost = node_map(true);
-
+  const auto& vmap_owned = ent_map_owned_[Entity_kind::NODE];
+  const auto& vmap_wghost = ent_map_wghost_[Entity_kind::NODE];
+  
   std::vector<int> gids;
-  Epetra_IntVector counts(fmap_wghost), flags(vmap_wghost);
+  Epetra_IntVector counts(*fmap_wghost.get()), flags(*vmap_wghost.get());
 
   // collect information about boundary faces and nodes
   for (int n = 0; n < nents; ++n) {
-    face_get_cells_internal_(n, Parallel_type::ALL, &cells);
+    getFaceCells(n, Parallel_type::ALL, cells);
     int ncells = cells.size();
     if (ncells == 1) {
-      gids.push_back(fmap_owned.GID(n));
+      gids.push_back(fmap_owned->GID(n));
 
-      face_get_nodes(n, &nodes);
+      getFaceNodes(n, nodes);
       for (int i = 0; i < nodes.size(); ++i) flags[nodes[i]] = 1;
     }
-    counts[n] = ncells;
+    counts[n] = ncells; 
   }
 
   // process faces
-  ent_extmap_owned_[FACE] = Teuchos::rcp(new Epetra_Map(-1, gids.size(), gids.data(), 0, *comm_));
-  exterior_face_importer_ =
-    Teuchos::rcp(new Epetra_Import(*ent_extmap_owned_[FACE], *ent_map_owned_[FACE]));
+  ent_extmap_owned_[Entity_kind::FACE] = Teuchos::rcp(new Epetra_Map(-1, gids.size(), gids.data(), 0, *comm_));
+  exterior_face_importer_ = Teuchos::rcp(new Epetra_Import(*ent_extmap_owned_[Entity_kind::FACE], *ent_map_owned_[Entity_kind::FACE]));
 
 #ifdef HAVE_MPI
   {
-    auto importer = Teuchos::rcp(new Epetra_Import(fmap_wghost, fmap_owned));
+    auto importer = Teuchos::rcp(new Epetra_Import(*fmap_wghost.get(), *fmap_owned.get()));
     int* vdata;
     counts.ExtractView(&vdata);
-    Epetra_IntVector tmp(View, fmap_owned, vdata);
+    Epetra_IntVector tmp(View, *fmap_owned.get(), vdata);
     counts.Import(tmp, *importer, Insert);
   }
 #endif
 
-  int nents_wghost = nents + nents_ghost_[FACE];
+  int nents_wghost = nents + nents_ghost_[Entity_kind::FACE];
   for (int n = nents; n < nents_wghost; ++n) {
-    if (counts[n] == 1) gids.push_back(fmap_wghost.GID(n));
+    if (counts[n] == 1) gids.push_back(fmap_wghost->GID(n));
   }
 
-  ent_extmap_wghost_[FACE] = Teuchos::rcp(new Epetra_Map(-1, gids.size(), gids.data(), 0, *comm_));
+  ent_extmap_wghost_[Entity_kind::FACE] = Teuchos::rcp(new Epetra_Map(-1, gids.size(), gids.data(), 0, *comm_));
 
   // process nodes
   gids.clear();
 
-  nents = nents_owned_[NODE];
+  nents = nents_owned_[Entity_kind::NODE];
   for (int n = 0; n < nents; ++n) {
-    if (flags[n] == 1) gids.push_back(vmap_owned.GID(n));
+    if (flags[n] == 1) gids.push_back(vmap_owned->GID(n));
   }
 
-  ent_extmap_owned_[NODE] = Teuchos::rcp(new Epetra_Map(-1, gids.size(), gids.data(), 0, *comm_));
+  ent_extmap_owned_[Entity_kind::NODE] = Teuchos::rcp(new Epetra_Map(-1, gids.size(), gids.data(), 0, *comm_));
 
 #ifdef HAVE_MPI
   {
-    auto importer = Teuchos::rcp(new Epetra_Import(vmap_wghost, vmap_owned));
+    auto importer = Teuchos::rcp(new Epetra_Import(*vmap_wghost.get(), *vmap_owned.get()));
     int* vdata;
     flags.ExtractView(&vdata);
-    Epetra_IntVector tmp(View, vmap_owned, vdata);
+    Epetra_IntVector tmp(View, *vmap_owned.get(), vdata);
     flags.Import(tmp, *importer, Insert);
   }
 #endif
 
-  nents_wghost = nents + nents_ghost_[NODE];
+  nents_wghost = nents + nents_ghost_[Entity_kind::NODE];
   for (int n = nents; n < nents_wghost; ++n) {
-    if (flags[n] == 1) gids.push_back(vmap_wghost.GID(n));
+    if (flags[n] == 1) gids.push_back(vmap_wghost->GID(n));
   }
 
-  ent_extmap_wghost_[NODE] = Teuchos::rcp(new Epetra_Map(-1, gids.size(), gids.data(), 0, *comm_));
+  ent_extmap_wghost_[Entity_kind::NODE] = Teuchos::rcp(new Epetra_Map(-1, gids.size(), gids.data(), 0, *comm_));
 }
 
 
 /* ******************************************************************
 * Exception due to limitations of the base mesh framework.
 ****************************************************************** */
-void
-MeshExtractedManifold::TryExtension_(const std::string& setname,
-                                     Entity_kind kind_p,
-                                     Entity_kind kind_d,
-                                     Entity_ID_List* setents) const
+void MeshExtractedManifold::TryExtension_(
+    const std::string& setname,
+    Entity_kind kind_p, Entity_kind kind_d, Entity_ID_List* setents) const
 {
   // labeled set: extract edges
   setents->clear();
 
-  const auto& gm = geometric_model();
+  const auto& gm = getGeometricModel();
   if (gm == Teuchos::null) return;
 
   auto rgn = gm->FindRegion(setname);
@@ -941,8 +956,8 @@ MeshExtractedManifold::TryExtension_(const std::string& setname,
   // populate list of edges
   std::vector<Entity_ID> faceents;
   std::vector<double> vofs;
-  parent_mesh_->get_set_entities_and_vofs(setname, FACE, Parallel_type::ALL, &faceents, &vofs);
-  auto marked_ents = EnforceOneLayerOfGhosts_(setname, FACE, &faceents);
+  std::tie(faceents,vofs) = getSetEntitiesAndVolumeFractions(setname, Entity_kind::FACE, Parallel_type::ALL);
+  auto marked_ents = EnforceOneLayerOfGhosts_(setname, Entity_kind::FACE, &faceents);
 
   Entity_ID_List edges, nodes;
   std::vector<int> dirs;
@@ -950,20 +965,27 @@ MeshExtractedManifold::TryExtension_(const std::string& setname,
 
   for (auto it = marked_ents.begin(); it != marked_ents.end(); ++it) {
     int f = it->first;
-    if (kind_p == FACE) {
+    if (kind_p == Entity_kind::FACE) {
       setents_tmp.insert(f);
-    } else if (kind_p == EDGE) {
-      parent_mesh_->face_get_edges_and_dirs(f, &edges, &dirs);
-      for (int i = 0; i < edges.size(); ++i) { setents_tmp.insert(edges[i]); }
-    } else if (kind_p == NODE) {
-      parent_mesh_->face_get_nodes(f, &nodes);
-      for (int i = 0; i < nodes.size(); ++i) { setents_tmp.insert(nodes[i]); }
+    }
+    else if (kind_p == Entity_kind::EDGE) {
+      parent_mesh_->getFaceEdgesAndDirs(f, edges, &dirs);
+      for (int i = 0; i < edges.size(); ++i) {
+        setents_tmp.insert(edges[i]);
+      }
+    }
+    else if (kind_p == Entity_kind::NODE) {
+      parent_mesh_->getFaceNodes(f, nodes);
+      for (int i = 0; i < nodes.size(); ++i) {
+        setents_tmp.insert(nodes[i]);
+      }
     }
   }
 
-  for (auto it = setents_tmp.begin(); it != setents_tmp.end(); ++it) setents->push_back(*it);
+  for (auto it = setents_tmp.begin(); it != setents_tmp.end(); ++it)
+    setents->push_back(*it);
 
-  std::string setname_internal = setname + std::to_string(kind_d);
+  std::string setname_internal = setname + std::to_string(static_cast<int>(kind_d));
   parent_labeledsets_[setname_internal] = *setents;
 }
 
@@ -971,40 +993,38 @@ MeshExtractedManifold::TryExtension_(const std::string& setname,
 /* ******************************************************************
 * Limits the set of parent objects to only one layer of ghosts.
 ****************************************************************** */
-std::map<Entity_ID, int>
-MeshExtractedManifold::EnforceOneLayerOfGhosts_(const std::string& setname,
-                                                Entity_kind kind,
-                                                Entity_ID_List* setents) const
+std::map<Entity_ID, int> MeshExtractedManifold::EnforceOneLayerOfGhosts_(
+    const std::string& setname, Entity_kind kind, Entity_ID_List* setents) const
 {
   // base set is the set of master cells
   Entity_ID_List fullset;
-  if (kind != FACE) {
+  if (kind != Entity_kind::FACE) {
     std::vector<double> vofs;
-    parent_mesh_->get_set_entities_and_vofs(setname, FACE, Parallel_type::ALL, &fullset, &vofs);
-  } else {
+    std::tie(fullset,vofs) = getSetEntitiesAndVolumeFractions(setname, Entity_kind::FACE, Parallel_type::ALL);
+  } else { 
     fullset = *setents;
   }
 
-  // initial set of entities is defined by master parent faces and is marked as
+  // initial set of entities is defined by master parent faces and is marked as 
   // potential master entities
   Entity_ID_List nodes, edges, faces;
   std::vector<int> dirs;
-  int nfaces_owned = parent_mesh_->num_entities(FACE, Parallel_type::OWNED);
+  int nfaces_owned = parent_mesh_->getNumEntities(Entity_kind::FACE, Parallel_type::OWNED);
   std::map<Entity_ID, int> nodeset0, nodeset, edgeset, faceset;
 
   for (int n = 0; n < fullset.size(); ++n) {
     int f = fullset[n];
     if (f < nfaces_owned) {
-      parent_mesh_->face_get_nodes(f, &nodes);
+      parent_mesh_->getFaceNodes(f, nodes);
       for (int i = 0; i < nodes.size(); ++i) nodeset0[nodes[i]] = MASTER;
-
-      if (kind == EDGE) {
-        parent_mesh_->face_get_edges_and_dirs(f, &edges, &dirs);
+ 
+      if (kind == Entity_kind::EDGE) {
+        parent_mesh_->getFaceEdgesAndDirs(f, edges, &dirs);
         for (int i = 0; i < edges.size(); ++i) edgeset[edges[i]] = MASTER;
       }
-
+ 
       faceset[f] = MASTER;
-    }
+    } 
   }
 
   // ghosts entities are defined by neighboor faces of master faces. New
@@ -1014,7 +1034,7 @@ MeshExtractedManifold::EnforceOneLayerOfGhosts_(const std::string& setname,
     int f = fullset[n];
     if (f >= nfaces_owned) {
       bool found(false);
-      parent_mesh_->face_get_nodes(f, &nodes);
+      parent_mesh_->getFaceNodes(f, nodes);
       for (int i = 0; i < nodes.size(); ++i) {
         if (nodeset0.find(nodes[i]) != nodeset0.end()) {
           found = true;
@@ -1030,8 +1050,8 @@ MeshExtractedManifold::EnforceOneLayerOfGhosts_(const std::string& setname,
             it->second |= GHOST;
         }
 
-        if (kind == EDGE) {
-          parent_mesh_->face_get_edges_and_dirs(f, &edges, &dirs);
+        if (kind == Entity_kind::EDGE) {
+          parent_mesh_->getFaceEdgesAndDirs(f, edges, &dirs);
           for (int i = 0; i < edges.size(); ++i) {
             auto it = edgeset.find(edges[i]);
             if (it == edgeset.end())
@@ -1050,17 +1070,17 @@ MeshExtractedManifold::EnforceOneLayerOfGhosts_(const std::string& setname,
   std::set<Entity_ID> auxset;
   for (int n = 0; n < fullset.size(); ++n) auxset.insert(fullset[n]);
 
-  if (kind == FACE) {
+  if (kind == Entity_kind::FACE) {
     return faceset;
-  } else if (kind == EDGE) {
-    const auto& fmap = parent_mesh_->face_map(true);
+  } else if (kind == Entity_kind::EDGE) {
+    const auto& fmap = parent_to_entid_[Entity_kind::FACE];
 
-    int nowned = parent_mesh_->num_entities(FACE, Parallel_type::OWNED);
-    int gidmax = fmap.MaxAllGID();
+    int nowned = parent_mesh_->getNumEntities(Entity_kind::FACE, Parallel_type::OWNED);
+    int gidmax = std::max_element(fmap.begin(), fmap.end())->second; 
 
     for (auto it = edgeset.begin(); it != edgeset.end(); ++it) {
       if (it->second == MASTER + GHOST) {
-        parent_mesh_->edge_get_faces(it->first, Parallel_type::ALL, &faces);
+        parent_mesh_->getEdgeFaces(it->first, Parallel_type::ALL, faces);
         int nfaces = faces.size();
 
         // compare maximum global ids for owned and all faces living on manifold
@@ -1068,23 +1088,24 @@ MeshExtractedManifold::EnforceOneLayerOfGhosts_(const std::string& setname,
         for (int n = 0; n < nfaces; ++n) {
           Entity_ID f = faces[n];
           if (auxset.find(f) != auxset.end()) {
-            gid_wghost_min = std::min(gid_wghost_min, fmap.GID(f));
-            if (f < nowned) gid_owned_min = std::min(gid_owned_min, fmap.GID(f));
+            gid_wghost_min = std::min(gid_wghost_min, fmap.at(f));
+            if (f < nowned)
+              gid_owned_min = std::min(gid_owned_min, fmap.at(f));
           }
         }
         it->second = (gid_wghost_min == gid_owned_min) ? MASTER : GHOST;
       }
     }
     return edgeset;
-  } else if (kind == NODE) {
-    const auto& fmap = parent_mesh_->face_map(true);
+  } else if (kind == Entity_kind::NODE) {
+    const auto& fmap = parent_to_entid_[Entity_kind::FACE];
 
-    int nowned = parent_mesh_->num_entities(FACE, Parallel_type::OWNED);
-    int gidmax = fmap.MaxAllGID();
+    int nowned = parent_mesh_->getNumEntities(Entity_kind::FACE, Parallel_type::OWNED);
+    int gidmax = std::max_element(fmap.begin(), fmap.end())->second; 
 
     for (auto it = nodeset.begin(); it != nodeset.end(); ++it) {
       if (it->second == MASTER + GHOST) {
-        parent_mesh_->node_get_faces(it->first, Parallel_type::ALL, &faces);
+        parent_mesh_->getNodeFaces(it->first, Parallel_type::ALL, faces);
         int nfaces = faces.size();
 
         // compare maximum global ids for owned and all faces living on manifold
@@ -1092,8 +1113,9 @@ MeshExtractedManifold::EnforceOneLayerOfGhosts_(const std::string& setname,
         for (int n = 0; n < nfaces; ++n) {
           Entity_ID f = faces[n];
           if (auxset.find(f) != auxset.end()) {
-            gid_wghost_min = std::min(gid_wghost_min, fmap.GID(f));
-            if (f < nowned) gid_owned_min = std::min(gid_owned_min, fmap.GID(f));
+            gid_wghost_min = std::min(gid_wghost_min, fmap.at(f));
+            if (f < nowned)
+              gid_owned_min = std::min(gid_owned_min, fmap.at(f));
           }
         }
         it->second = (gid_wghost_min == gid_owned_min) ? MASTER : GHOST;
@@ -1107,55 +1129,30 @@ MeshExtractedManifold::EnforceOneLayerOfGhosts_(const std::string& setname,
 
 
 /* ******************************************************************
-* Check if point is in 3D polygon by Jordan's crossing algorithm
-****************************************************************** */
-bool
-MeshExtractedManifold::PointInCell_(const AmanziGeometry::Point& xp, int c) const
-{
-  Entity_ID_List nodes;
-  cell_get_nodes(c, &nodes);
-  int nnodes = nodes.size();
-
-  double vol(0.0);
-  AmanziGeometry::Point xp1(3), xp2(3);
-
-  for (int i1 = 0; i1 < nnodes; ++i1) {
-    int i2 = (i1 + 1) % nnodes;
-    node_get_coordinates(nodes[i1], &xp1);
-    node_get_coordinates(nodes[i2], &xp2);
-    vol += norm((xp1 - xp) ^ (xp2 - xp));
-  }
-
-  double area = cell_volume(c);
-  return (fabs(vol / 2 - area) < 1e-14 * area);
-}
-
-
-/* ******************************************************************
 * Statistics
 ****************************************************************** */
-void
-MeshExtractedManifold::PrintSets_() const
+void MeshExtractedManifold::PrintSets_() const
 {
   if (vo_.get() && vo_->os_OK(Teuchos::VERB_LOW)) {
     Teuchos::OSTab tab = vo_->getOSTab();
     *(vo_->os()) << "sets: ";
     for (auto it : sets_) {
       int i1, i0(it.second.size());
-      get_comm()->SumAll(&i0, &i1, 1);
+      getComm()->SumAll(&i0, &i1, 1);
       *(vo_->os()) << "\"" << it.first << "\" (" << i1 << ") ";
     }
-    *(vo_->os()) << "\n";
+    *(vo_->os())  << "\n";
 
     *(vo_->os()) << "parent labeledsets: ";
     for (auto it : parent_labeledsets_) {
       int i1, i0(it.second.size());
-      get_comm()->SumAll(&i0, &i1, 1);
+      getComm()->SumAll(&i0, &i1, 1);
       *(vo_->os()) << "\"" << it.first << "\" (" << i1 << ") ";
     }
-    *(vo_->os()) << "\n";
+    *(vo_->os())  << "\n";
   }
 }
 
-} // namespace AmanziMesh
-} // namespace Amanzi
+}  // namespace AmanziMesh
+}  // namespace Amanzi
+
