@@ -18,21 +18,19 @@ namespace Amanzi {
 namespace AmanziMesh {
 
 // Identifers for those file formats understood
-enum struct FileFormat {
-  UNKNOWN = 0, /* It's a mystery */
-  EXODUS_II,   /* Exodus II format */
-  NEMESIS,     /* Exodus II format partitioned by Nemesis */
-  MOAB_HDF5    /* HDF5 format used by MOAB */
+enum struct FileFormat { 
+  UNKNOWN = 0,          /* It's a mystery */
+  EXODUS_II,            /* Exodus II format */
+  NEMESIS,              /* Exodus II format partitioned by Nemesis */
+  MOAB_HDF5             /* HDF5 format used by MOAB */
 };
 
 // Get the name of a particular file format
-std::string
-fileFormatName(const FileFormat f);
+std::string fileFormatName(const FileFormat f);
 
 // Determine a file format from a file name, and perform sanity checks on that
 // filename.
-FileFormat
-fileFormatFromFilename(const Comm_type& comm_, std::string name);
+FileFormat fileFormatFromFilename(const Comm_type& comm_, std::string name);
 
 } // namespace AmanziMesh
 } // namespace Amanzi
