@@ -20,8 +20,8 @@ namespace AmanziMesh {
 /// A special exception type for MeshFactory errors
 class Message : public Errors::Message {
  public:
-  explicit Message(void) : Errors::Message(){};
-  explicit Message(const char* message) : Errors::Message(message){};
+  explicit Message(void) : Errors::Message() {};
+  explicit Message(const char* message) : Errors::Message(message) {};
   virtual ~Message(void) noexcept {};
 };
 
@@ -29,20 +29,20 @@ class Message : public Errors::Message {
 // A special exception to identify file problems
 class FileMessage : public Message {
  public:
-  explicit FileMessage(void) : Message(){};
-  explicit FileMessage(const char* message) : Message(message){};
+  explicit FileMessage(void) : Message() {};
+  explicit FileMessage(const char* message) : Message(message) {};
   virtual ~FileMessage(void) noexcept {};
-};
+};  
 
 // A special exception to identify frameworks problems
 class FrameworkMessage : public Message {
  public:
-  explicit FrameworkMessage(void) : Message(){};
-  explicit FrameworkMessage(const char* message) : Message(message){};
+  explicit FrameworkMessage(void) : Message() {};
+  explicit FrameworkMessage(const char* message) : Message(message) {};
   virtual ~FrameworkMessage(void) noexcept {};
-};
+};  
 
-} // namespace AmanziMesh
-} // namespace Amanzi
+}  // namespace AmanziMesh
+}  // namespace Amanzi
 
 #endif
