@@ -8,37 +8,34 @@
   Authors: William Perkins, others
 */
 
-#ifndef AMANZI_MESH_FRAMEWORK_HH_
-#define AMANZI_MESH_FRAMEWORK_HH_
+#pragma once
 
 #include <string>
 #include <vector>
-#include <map>
 
 namespace Amanzi {
 namespace AmanziMesh {
 
-enum struct Framework { SIMPLE = 0, MSTK, MOAB, STK };
+enum struct Framework { SIMPLE=0, MSTK, MOAB };
 
 using Preference = std::vector<Framework>;
 
-static const std::map<Framework, std::string> framework_names = {
-  { Framework::MSTK, std::string("MSTK") },
-  { Framework::MOAB, std::string("MOAB") },
-  { Framework::STK, std::string("stk:mesh") },
-  { Framework::SIMPLE, std::string("Simple") }
-};
 
+inline
+std::string to_string(const Framework framework) {
+  switch(framework) {
+    case(Framework::MSTK): return "MSTK";
+    case(Framework::SIMPLE): return "Simple";
+    case(Framework::MOAB): return "MOAB";
+    default: return "unknown";
+  }
+}
 
-Preference
-default_preference();
-bool
-framework_enabled(Framework f);
-Preference
-filter_preference(const Preference& pref);
+Preference default_preference();
+bool framework_enabled(Framework f);
+Preference filter_preference(const Preference& pref);
 
 
 } // namespace AmanziMesh
 } // namespace Amanzi
 
-#endif
