@@ -3,11 +3,14 @@
 #include "Teuchos_GlobalMPISession.hpp"
 
 #include "VerboseObject_objs.hh"
+#include "Kokkos_Core.hpp"
 
-int
-main(int argc, char* argv[])
+int main(int argc, char *argv[])
 {
-  Teuchos::GlobalMPISession mpiSession(&argc, &argv);
-
-  return UnitTest::RunAllTests();
+  Teuchos::GlobalMPISession mpiSession(&argc,&argv);
+  Kokkos::initialize(); 
+  auto r =  UnitTest::RunAllTests ();
+  Kokkos::finalize(); 
+  return r; 
 }
+
