@@ -23,85 +23,83 @@
 // check_preference
 // -------------------------------------------------------------
 
-SUITE(MeshFramework)
-{
-  TEST(EXTRACT_COMM_SPLIT)
-  {
-    auto comm = Amanzi::getDefaultComm();
+SUITE (MeshFramework) {
 
-    double x0(0), y0(0), z0(0);
-    double x1(10), y1(10), z1(10);
-    int nx(10), ny(10), nz(10);
+TEST (EXTRACT_COMM_SPLIT) {
+  auto comm = Amanzi::getDefaultComm();
 
-    // make regions
-    Teuchos::ParameterList parameterlist;
-    Teuchos::ParameterList& reg_spec = parameterlist.sublist("regions");
-    Teuchos::ParameterList& top_surface = reg_spec.sublist("Top Surface");
-    Teuchos::ParameterList& top_surface_def = top_surface.sublist("region: box");
-    Teuchos::Array<double> low = Teuchos::tuple(0.0, 0.0, 10.0);
-    Teuchos::Array<double> high = Teuchos::tuple(10.0, 10.0, 10.0);
-    top_surface_def.set<Teuchos::Array<double>>("low coordinate", low);
-    top_surface_def.set<Teuchos::Array<double>>("high coordinate", high);
+  double x0(0), y0(0), z0(0);
+  double x1(10), y1(10), z1(10);
+  int nx(10), ny(10), nz(10);
 
-    Teuchos::RCP<Amanzi::AmanziGeometry::GeometricModel> gm =
-      Teuchos::rcp(new Amanzi::AmanziGeometry::GeometricModel(3, reg_spec, *comm));
+  // make regions
+  Teuchos::ParameterList parameterlist;
+  Teuchos::ParameterList& reg_spec = parameterlist.sublist("regions");
+  Teuchos::ParameterList& top_surface = reg_spec.sublist("Top Surface");
+  Teuchos::ParameterList& top_surface_def = top_surface.sublist("region: box");
+  Teuchos::Array<double> low = Teuchos::tuple(0.0,0.0,10.0);
+  Teuchos::Array<double> high = Teuchos::tuple(10.0,10.0,10.0);
+  top_surface_def.set< Teuchos::Array<double> >("low coordinate",low);
+  top_surface_def.set< Teuchos::Array<double> >("high coordinate",high);
 
-    std::vector<std::string> setnames{ "Top Surface" };
+  Teuchos::RCP<Amanzi::AmanziGeometry::GeometricModel> gm =
+    Teuchos::rcp(new Amanzi::AmanziGeometry::GeometricModel(3, reg_spec, *comm));
 
-    Amanzi::AmanziMesh::Preference pref;
+  std::vector<std::string> setnames{"Top Surface"};
 
-    auto plist = Teuchos::rcp(new Teuchos::ParameterList());
-    plist->sublist("unstructured").sublist("expert").set<std::string>("partitioner", "metis");
-    plist->sublist("unstructured").sublist("expert").set<bool>("create subcommunicator", true);
-    Amanzi::AmanziMesh::MeshFactory meshfactory(comm, gm, plist);
+  Amanzi::AmanziMesh::Preference pref;
 
-    bool flatten = true;
-    // bool extrude = false;
+  auto plist = Teuchos::rcp(new Teuchos::ParameterList());
+  plist->sublist("unstructured").sublist("expert").set<std::string>("partitioner", "metis");
+  plist->sublist("unstructured").sublist("expert").set<bool>("create subcommunicator", true);
+  Amanzi::AmanziMesh::MeshFactory meshfactory(comm,gm,plist);
 
-    if (Amanzi::AmanziMesh::framework_enabled(Amanzi::AmanziMesh::Framework::MSTK)) {
-      pref.clear();
-      pref.push_back(Amanzi::AmanziMesh::Framework::MSTK);
-      meshfactory.set_preference(pref);
+  bool flatten = true;
+  bool extrude = false;
 
-      auto parent_mesh = meshfactory.create(x0, y0, z0, x1, y1, z1, nx, ny, nz);
-      CHECK(!parent_mesh.is_null());
+  if (Amanzi::AmanziMesh::framework_enabled(
+        Amanzi::AmanziMesh::Framework::MSTK)) {
 
-      Amanzi::AmanziMesh::Entity_ID_List setents;
-      parent_mesh->get_set_entities("Top Surface",
-                                    Amanzi::AmanziMesh::FACE,
-                                    Amanzi::AmanziMesh::Parallel_type::OWNED,
-                                    &setents);
-      int local_count = setents.size();
-      int global_min_count;
-      std::cout << "Count on rank " << comm->MyPID() << " is " << local_count << std::endl;
-      comm->MinAll(&local_count, &global_min_count, 1);
-      AMANZI_ASSERT(global_min_count == 0); // this test only useful for min = 0
+    pref.clear(); pref.push_back(Amanzi::AmanziMesh::Framework::MSTK);
+    meshfactory.set_preference(pref);
 
-      int local_cells = parent_mesh->num_entities(Amanzi::AmanziMesh::CELL,
-                                                  Amanzi::AmanziMesh::Parallel_type::OWNED);
-      int global_cells;
-      comm->SumAll(&local_cells, &global_cells, 1);
-      CHECK(global_cells == 1000);
+    auto parent_mesh = meshfactory.create(x0, y0, z0,
+            x1, y1, z1,
+            nx, ny, nz);
+    CHECK(!parent_mesh.is_null());
 
-      // BELOW SEGMENT DEMONSTRATES MSTK ISSUE #112
-      int max_cell = -1;
-      for (int c = 0; c != local_cells; ++c) {
-        max_cell = std::max(max_cell, parent_mesh->cell_map(false).GID(c));
-      }
-      int global_max_cell;
-      comm->MaxAll(&max_cell, &global_max_cell, 1);
-      std::cout << "cell counts: total = " << global_cells << ", max = " << global_max_cell
-                << std::endl;
-      // REMOVE COMMENT ONCE MSTK IS UPDATED
-      //CHECK(global_max_cell == 999);
-      // END SEGMENT
+    Amanzi::AmanziMesh::Entity_ID_List setents = parent_mesh->getSetEntities(
+      "Top Surface", Amanzi::AmanziMesh::Entity_kind::FACE, Amanzi::AmanziMesh::Parallel_type::OWNED);
+    int local_count = setents.size();
+    int global_min_count;
+    std::cout << "Count on rank " << comm->MyPID() << " is " << local_count << std::endl;
+    comm->MinAll(&local_count, &global_min_count, 1);
+    AMANZI_ASSERT(global_min_count == 0);  // this test only useful for min = 0
 
-      auto newmesh = meshfactory.create(parent_mesh, setnames, Amanzi::AmanziMesh::FACE, flatten);
-      if (local_count == 0) {
-        CHECK(newmesh == Teuchos::null);
-      } else {
-        CHECK(newmesh != Teuchos::null);
-      }
+    int local_cells = parent_mesh->getNumEntities(Amanzi::AmanziMesh::Entity_kind::CELL, Amanzi::AmanziMesh::Parallel_type::OWNED);
+    int global_cells;
+    comm->SumAll(&local_cells, &global_cells, 1);
+    CHECK(global_cells == 1000);
+
+    // BELOW SEGMENT DEMONSTRATES MSTK ISSUE #112
+    int max_cell = -1;
+    for (int c=0; c!=local_cells; ++c) {
+      max_cell = std::max(max_cell, parent_mesh->getMap(Amanzi::AmanziMesh::Entity_kind::CELL,false).GID(c));
     }
+    int global_max_cell;
+    comm->MaxAll(&max_cell, &global_max_cell, 1);
+    std::cout << "cell counts: total = " << global_cells << ", max = " << global_max_cell << std::endl;
+    // REMOVE COMMENT ONCE MSTK IS UPDATED
+    CHECK(global_max_cell == 999);
+    // END SEGMENT
+
+    // Which constructor? 
+    //auto newmesh = meshfactory.create(parent_mesh, setnames, Amanzi::AmanziMesh::Entity_kind::FACE, flatten);
+    //if (local_count == 0) {
+    //  CHECK(newmesh == Teuchos::null);
+    //} else {
+    //  CHECK(newmesh != Teuchos::null);
+    //}
   }
+}
 }
