@@ -1,718 +1,392 @@
 #include "Epetra_IntVector.h"
 
 #include "RegionEnumerated.hh"
+#include "MeshLogical.hh"
 #include "MeshEmbeddedLogical.hh"
 
 namespace Amanzi {
 namespace AmanziMesh {
 
-//
-// MeshEmbeddedLogical Constructor
-//
-// Combines two meshes, each representing their own domain, into a
-// single mesh via set of specified connections.  The result is a
-// logical mesh in the sense that it provides a limited interface.
-//
-//  - face_cell_list          : length nfaces array of length 2 arrays
-//                              defining the topology
-//  - face_cell_lengths       : length of the cell-to-face connection
-//  - face_area_normals       : length nfaces array of normals of the
-//                              face, points from cell 1 to 2 in
-//                              face_cell_list topology, magnitude
-//                              is area
-MeshEmbeddedLogical::MeshEmbeddedLogical(
-  const Comm_ptr_type& comm,
-  Teuchos::RCP<Mesh> bg_mesh,
-  Teuchos::RCP<Mesh> log_mesh,
-  const std::vector<Entity_ID_List>& face_cell_ids,
-  const std::vector<std::vector<double>>& face_cell_lengths,
-  const std::vector<AmanziGeometry::Point>& face_area_normals,
-  const Teuchos::RCP<const Teuchos::ParameterList>& plist)
-  : Mesh(comm, Teuchos::null, plist, true, false), bg_mesh_(bg_mesh), log_mesh_(log_mesh)
+/*
+MeshEmbeddedLogical Constructor
+
+Combines two meshes, each representing their own domain, into a
+single mesh via set of specified connections.  The result is a
+logical mesh in the sense that it provides a limited interface.
+
+ - face_cell_list          : length nfaces array of length 2 arrays
+                             defining the topology
+ - face_cell_lengths       : length of the cell-to-face connection
+ - face_area_normals       : length nfaces array of normals of the
+                             face, points from cell 1 to 2 in
+                             face_cell_list topology, magnitude
+                             is area
+
+Note that entities are ordered in the following way:
+
+1. logical mesh entities
+2. "extra" entities (valid only for faces, these are
+   the connections between the two meshes.
+3. background entities
+
+*/
+MeshEmbeddedLogical::MeshEmbeddedLogical(const Comm_ptr_type& comm,
+        Teuchos::RCP<MeshFramework> bg_mesh,
+        Teuchos::RCP<MeshFramework> log_mesh,
+        const std::vector<Entity_ID_List>& face_cell_ids,
+        const std::vector<std::vector<double> >& face_cell_lengths,
+        const std::vector<AmanziGeometry::Point>& face_area_normals,
+        const Teuchos::RCP<Teuchos::ParameterList>& plist)
+    : MeshFramework(comm, Teuchos::null, plist),
+      bg_mesh_(bg_mesh),
+      log_mesh_(log_mesh),
+      extra_face_cell_ids_(face_cell_ids),
+      extra_face_cell_lengths_(face_cell_lengths),
+      extra_face_area_normals_(face_area_normals)
 {
-  set_space_dimension(3);
-  set_manifold_dimension(3); // mixed?
+  setSpaceDimension(3);
+  setManifoldDimension(3);
+  setAlgorithms(Teuchos::rcp(new MeshLogicalAlgorithms()));
 
-  // ensure cached properties are available
-  if (!bg_mesh->cell_geometry_precomputed_) bg_mesh->compute_cell_geometric_quantities_();
-  if (!bg_mesh->face_geometry_precomputed_) bg_mesh->compute_face_geometric_quantities_();
-  if (!bg_mesh->cell2face_info_cached_ || !bg_mesh->face2cell_info_cached_)
-    bg_mesh->cache_cell2face_info_();
+  // merge and remap to get new global IDs
+  int ncells_bg_owned = bg_mesh->getNumEntities(Entity_kind::CELL, Parallel_type::OWNED);
+  int ncells_bg_all = bg_mesh->getNumEntities(Entity_kind::CELL, Parallel_type::ALL);
+  int ncells_log = log_mesh->getNumEntities(Entity_kind::CELL, Parallel_type::OWNED);
+  int ncells_my_all = ncells_bg_all + ncells_log;
 
-  // ensure cached properties are available
-  if (!log_mesh->cell_geometry_precomputed_) log_mesh->compute_cell_geometric_quantities_();
-  if (!log_mesh->face_geometry_precomputed_) log_mesh->compute_face_geometric_quantities_();
-  if (!log_mesh->cell2face_info_cached_ || !log_mesh->face2cell_info_cached_)
-    log_mesh->cache_cell2face_info_();
+  int nfaces_bg_owned = bg_mesh->getNumEntities(Entity_kind::FACE, Parallel_type::OWNED);
+  int nfaces_bg_all = bg_mesh->getNumEntities(Entity_kind::FACE, Parallel_type::ALL);
+  int nfaces_log = log_mesh->getNumEntities(Entity_kind::FACE, Parallel_type::OWNED);
+  int nfaces_extra = extra_face_cell_ids_.size<MemSpace_type::HOST>();
 
-  // merge and remap
-  int ncells_bg_owned = bg_mesh->num_entities(CELL, Parallel_type::OWNED);
-  int ncells_bg_used = bg_mesh->num_entities(CELL, Parallel_type::ALL);
-  int ncells_log = log_mesh->num_entities(CELL, Parallel_type::OWNED);
-  // int ncells_my_owned = ncells_bg_owned + ncells_log;
-  int ncells_my_used = ncells_bg_used + ncells_log;
-
-  int nfaces_bg_owned = bg_mesh->num_entities(FACE, Parallel_type::OWNED);
-  int nfaces_bg_used = bg_mesh->num_entities(FACE, Parallel_type::ALL);
-  int nfaces_log = log_mesh->num_entities(FACE, Parallel_type::OWNED);
-  int nfaces_extra = face_cell_ids.size();
-  // int nfaces_my_owned = nfaces_bg_owned + nfaces_log + nfaces_extra;
-  // int nfaces_my_used = nfaces_bg_used + nfaces_log + nfaces_extra;
-
-  //
-  // face ordered things first
-  // ----------------------------
-  //
-  // Note that new face lids are ordered:
-  //  1. logical-logical connections
-  //  2. logical-background connections (these are the new ones)
-  //  3. background-background connections
-  //
-  // face-to-cell map -- remap and add in extras
-  // -- first faces are the logical-logical faces
-  face_cell_ids_ = log_mesh->face_cell_ids_;
-  // -- next are the logical-background faces, insert and remap.
-  //
-  // Per stated convention of normal direction, take 1s complement of
-  // second cell (if not done already)
-  std::vector<Entity_ID_List> face_cell_ids_copy = face_cell_ids;
-  for (auto& fc : face_cell_ids_copy)
-    if (!std::signbit(fc[1])) fc[1] = ~fc[1];
-  face_cell_ids_.insert(face_cell_ids_.end(), face_cell_ids_copy.begin(), face_cell_ids_copy.end());
-  for (int f = nfaces_log; f != face_cell_ids_.size(); ++f) {
-    // all new faces are logical,bg.  must remap the bg cell. since
-    // cell 1 is storing the 1s complement of the ID, we have convert
-    // it to a regular ID by taking the 1s complement, offset it and
-    // then convert it back
-    face_cell_ids_[f][1] = ~(~face_cell_ids_[f][1] + ncells_log);
-  }
-  // -- finally the bg-bg faces, insert and remap
-  face_cell_ids_.insert(
-    face_cell_ids_.end(), bg_mesh->face_cell_ids_.begin(), bg_mesh->face_cell_ids_.end());
-  for (int f = nfaces_log + nfaces_extra; f != face_cell_ids_.size(); ++f) {
-    int n_cells = face_cell_ids_[f].size();
-    for (int i = 0; i != n_cells; ++i)
-      if (std::signbit(face_cell_ids_[f][i]))
-        face_cell_ids_[f][i] = ~(~face_cell_ids_[f][i] + ncells_log);
-      else
-        face_cell_ids_[f][i] += ncells_log;
-  }
-
-  // // face normals -- same order: log-log, log-bg, bg-bg
-  // face_normal0_ = log_mesh->face_normal0_;
-  // face_normal0_.insert(face_normal0_.end(), face_area_normals.begin(),
-  //                      face_area_normals.end());
-  // face_normal0_.insert(face_normal0_.end(), bg_mesh->face_normal0_.begin(),
-  //                      bg_mesh->face_normal0_.end());
-
-  // face_normal1_ = log_mesh->face_normal1_;
-  // face_normal1_.insert(face_normal1_.end(), face_area_normals.begin(),
-  //                      face_area_normals.end());
-  // // -- negate normal1, normal direction implied from log->bg
-  // for (int f=nfaces_log; f!=face_normal1_.size(); ++f) {
-  //   face_normal1_[f] = -face_normal1_[f];
-  // }
-  // face_normal1_.insert(face_normal1_.end(), bg_mesh->face_normal1_.begin(),
-  //                      bg_mesh->face_normal1_.end());
-
-  face_normals_ = log_mesh->face_normals_;
-  for (int i = 0; i < face_area_normals.size(); i++)
-    face_normals_.emplace_back(
-      std::vector<AmanziGeometry::Point>({ face_area_normals[i], -face_area_normals[i] }));
-
-  face_normals_.insert(
-    face_normals_.end(), bg_mesh->face_normals_.begin(), bg_mesh->face_normals_.end());
-
-
-  // ptypes -- no remap.  all added faces are owned-owned
-  face_cell_ptype_ = log_mesh->face_cell_ptype_;
-  std::vector<Parallel_type> extra_ptypes(2, Parallel_type::OWNED);
-  for (int f = 0; f != nfaces_extra; ++f) { face_cell_ptype_.push_back(extra_ptypes); }
-  face_cell_ptype_.insert(
-    face_cell_ptype_.end(), bg_mesh->face_cell_ptype_.begin(), bg_mesh->face_cell_ptype_.end());
-
-  // areas -- no remap
-  face_areas_ = log_mesh->face_areas_;
-  for (int f = nfaces_log; f != nfaces_log + nfaces_extra; ++f) {
-    face_areas_.push_back(AmanziGeometry::norm(face_normals_[f][0]));
-  }
-  face_areas_.insert(face_areas_.end(), bg_mesh->face_areas_.begin(), bg_mesh->face_areas_.end());
-
-  //
-  // cell-ordered
-  // ----------------------------
-  //
-  // cell-to-face map -- remap
-  cell_face_ids_ = log_mesh->cell_face_ids_;
-  cell_face_ids_.insert(
-    cell_face_ids_.end(), bg_mesh->cell_face_ids_.begin(), bg_mesh->cell_face_ids_.end());
-  AMANZI_ASSERT(cell_face_ids_.size() == ncells_my_used);
-  for (int c = ncells_log; c != cell_face_ids_.size(); ++c) {
-    int n_faces = cell_face_ids_[c].size();
-    for (int i = 0; i != n_faces; ++i) { cell_face_ids_[c][i] += nfaces_log + nfaces_extra; }
-  }
-
-  // directions
-  cell_face_dirs_ = log_mesh->cell_face_dirs_;
-  cell_face_dirs_.insert(
-    cell_face_dirs_.end(), bg_mesh->cell_face_dirs_.begin(), bg_mesh->cell_face_dirs_.end());
-
-  // bisectors -- special interface since the cache isn't used for Mesh
-  Entity_ID_List faces;
-  cell_face_bisectors_.resize(cell_face_ids_.size());
-  for (int c = 0; c != ncells_log; ++c) {
-    log_mesh_->cell_get_faces_and_bisectors(c, &faces, &cell_face_bisectors_[c]);
-  }
-  for (int c = 0; c != ncells_bg_used; ++c) {
-    bg_mesh_->cell_get_faces_and_bisectors(c, &faces, &cell_face_bisectors_[c + ncells_log]);
-  }
+  // compute "extra" bisectors
+  extra_face_cell_bisectors_.resize(nfaces_extra, 2);
 
   // now loop over new faces, adding the updates to the cell-ordered versions
-  for (int f = nfaces_log; f != nfaces_log + nfaces_extra; ++f) {
-    Entity_ID c0 = face_cell_ids_[f][0];
-    Entity_ID c1 = ~face_cell_ids_[f][1];
-    cell_face_ids_[c0].push_back(f);
-    cell_face_ids_[c1].push_back(f);
-    cell_face_dirs_[c0].push_back(1);
-    cell_face_dirs_[c1].push_back(-1);
-    cell_face_bisectors_[c0].push_back(face_cell_lengths[f - nfaces_log][0] /
-                                       AmanziGeometry::norm(face_normals_[f][0]) *
-                                       face_normals_[f][0]);
-    cell_face_bisectors_[c1].push_back(face_cell_lengths[f - nfaces_log][1] /
-                                       AmanziGeometry::norm(face_normals_[f][1]) *
-                                       face_normals_[f][1]);
+  for (int fi=0; fi!=nfaces_extra; ++fi) {
+    auto normal0 = extra_face_area_normals_[fi];
+    extra_face_cell_bisectors_.getRow<MemSpace_type::HOST>(fi)[0] = face_cell_lengths[fi][0] / AmanziGeometry::norm(normal0) * normal0;
+
+    auto normal1 = extra_face_area_normals_[fi];
+    extra_face_cell_bisectors_.getRow<MemSpace_type::HOST>(fi)[1] = face_cell_lengths[fi][1] / AmanziGeometry::norm(normal1) * normal1;
   }
+  extra_face_cell_bisectors_.update<MemSpace_type::DEVICE>(); 
 
-  // cell volumes, just a straightforward merge
-  cell_volumes_ = log_mesh->cell_volumes_;
-  cell_volumes_.insert(
-    cell_volumes_.end(), bg_mesh->cell_volumes_.begin(), bg_mesh->cell_volumes_.end());
+  // Need to renumber the global IDs.  Use the corresponding mesh to communicate.
+  // First cells
+  // -- create a map of owned CELLs
+  int ncells_total_owned = getNumEntities(Entity_kind::CELL, Parallel_type::OWNED);
+  int ncells_total_all = getNumEntities(Entity_kind::CELL, Parallel_type::ALL);
+  Epetra_Map cell_owned_map(-1, ncells_total_owned, 0, *getComm());
 
-  //
-  // centroids
-  // ---------------------
-  if (bg_mesh->cell_centroids_.size() > 0) {
-    // -- cell, straighforward merge
-    cell_centroids_ = log_mesh->cell_centroids_;
-    cell_centroids_.insert(
-      cell_centroids_.end(), bg_mesh->cell_centroids_.begin(), bg_mesh->cell_centroids_.end());
+  // -- create a map on the background mesh of all CELLs
+  Entity_GID_List bg_cell_gids = bg_mesh_->getEntityGIDs(Entity_kind::CELL, Parallel_type::ALL);
+  Epetra_Map bg_cell_all_map(-1, ncells_bg_all, bg_cell_gids.data(), 0, *bg_mesh_->getComm());
+  Epetra_Map bg_cell_owned_map(-1, ncells_bg_owned, bg_cell_gids.data(), 0, *bg_mesh_->getComm());
 
-    // -- face, take new face centroids as mean of cell centroids
-    face_centroids_ = log_mesh->face_centroids_;
-    for (int f = nfaces_log; f != nfaces_log + nfaces_extra; ++f) {
-      Entity_ID c0 = face_cell_ids_[f][0];
-      Entity_ID c1 = ~face_cell_ids_[f][1];
-      face_centroids_.push_back((cell_centroids_[c0] + cell_centroids_[c1]) / 2.0);
-    }
-    face_centroids_.insert(
-      face_centroids_.end(), bg_mesh->face_centroids_.begin(), bg_mesh->face_centroids_.end());
-  }
+  // -- create a vector, fill the owned entries, and scatter
+  Epetra_IntVector bg_cell_owned_vec(bg_cell_owned_map);
+  for (int c=0; c!=ncells_bg_owned; ++c) bg_cell_owned_vec[c] = cell_owned_map.GID(c+ncells_log);
+  Epetra_IntVector bg_cell_all_vec(bg_cell_all_map);
+  Epetra_Import cell_import(bg_cell_all_map, bg_cell_owned_map);
+  bg_cell_all_vec.Import(bg_cell_owned_vec, cell_import, Insert);
 
-  // set up counts
-  num_entities_owned_[CELL] = ncells_log + ncells_bg_owned;
-  num_entities_owned_[FACE] = nfaces_log + nfaces_extra + nfaces_bg_owned;
-  num_entities_owned_[BOUNDARY_FACE] = bg_mesh->num_entities(BOUNDARY_FACE, Parallel_type::OWNED) +
-                                       log_mesh->num_entities(BOUNDARY_FACE, Parallel_type::OWNED);
-  num_entities_owned_[NODE] = 0;
+  // -- copy ghost GIDs into the renumbered vector
+  Entity_GID_List cell_gids(ncells_bg_all + ncells_log);
+  for (int c=0; c!=ncells_total_owned; ++c) cell_gids[c] = cell_owned_map.GID(c);
+  for (int c=ncells_total_owned; c!=ncells_total_all; ++c) cell_gids[c] = bg_cell_all_vec[c-ncells_log+ncells_bg_owned];
 
-  num_entities_used_[CELL] = ncells_log + ncells_bg_used;
-  num_entities_used_[FACE] = nfaces_log + nfaces_extra + nfaces_bg_used;
-  num_entities_used_[BOUNDARY_FACE] = bg_mesh->num_entities(BOUNDARY_FACE, Parallel_type::ALL) +
-                                      log_mesh->num_entities(BOUNDARY_FACE, Parallel_type::ALL);
-  num_entities_used_[NODE] = 0;
+  // -- create the map
+  cell_map_ = Teuchos::rcp(new Epetra_Map(-1, cell_gids.size(), cell_gids.data(), 0, *getComm()));
 
-  // toggle flags
-  cell_geometry_precomputed_ = true;
-  face_geometry_precomputed_ = true;
-  cell2face_info_cached_ = true;
-  faces_requested_ = true;
-  face2cell_info_cached_ = true;
+  // Next faces
+  // -- create a map of owned FACES
+  int nfaces_total_owned = getNumEntities(Entity_kind::FACE, Parallel_type::OWNED);
+  int nfaces_total_all = getNumEntities(Entity_kind::FACE, Parallel_type::ALL);
+  Epetra_Map face_owned_map(-1, nfaces_total_owned, 0, *getComm());
 
-  // build epetra maps
-  init_maps();
+  // -- create a map on the background mesh of all FACEs
+  Entity_GID_List bg_face_gids = bg_mesh_->getEntityGIDs(Entity_kind::FACE, Parallel_type::ALL);
+  Epetra_Map bg_face_all_map(-1, nfaces_bg_all, bg_face_gids.data(), 0, *bg_mesh_->getComm());
+  Epetra_Map bg_face_owned_map(-1, nfaces_bg_owned, bg_face_gids.data(), 0, *bg_mesh_->getComm());
+
+  // -- create a vector, fill the owned entries, and scatter
+  Epetra_IntVector bg_face_owned_vec(bg_face_owned_map);
+  for (int f=0; f!=nfaces_bg_owned; ++f) bg_face_owned_vec[f] = face_owned_map.GID(f+nfaces_log+nfaces_extra);
+  Epetra_IntVector bg_face_all_vec(bg_face_all_map);
+  Epetra_Import face_import(bg_face_all_map, bg_face_owned_map);
+  bg_face_all_vec.Import(bg_face_owned_vec, face_import, Insert);
+
+  // -- copy ghost GIDs into the renumbered vector
+  Entity_GID_List face_gids(nfaces_total_owned + nfaces_total_all);
+  for (int f=0; f!=nfaces_total_owned; ++f) face_gids[f] = face_owned_map.GID(f);
+  for (int f=nfaces_total_owned; f!=nfaces_total_all; ++f)
+    face_gids[f] = bg_face_all_vec[f-nfaces_log-nfaces_extra+nfaces_bg_owned];
+
+  // -- create the map
+  face_map_ = Teuchos::rcp(new Epetra_Map(-1, face_gids.size(), face_gids.data(), 0, *getComm()));
 }
 
-
-// build maps
-void
-MeshEmbeddedLogical::init_maps()
-{
-  // owned maps
-  // -- cell map
-  maps_owned_[CELL] = Teuchos::rcp(new Epetra_Map(-1, num_entities_owned_[CELL], 0, *comm_));
-
-  // -- face map
-  Teuchos::RCP<Epetra_Map> face_map =
-    Teuchos::rcp(new Epetra_Map(-1, num_entities_owned_[FACE], 0, *comm_));
-  maps_owned_[FACE] = face_map;
-
-  // exterior face map and importer
-  std::vector<int> extface_ids;
-  int nfaces_owned = num_entities_owned_[FACE];
-  for (int f = 0; f != nfaces_owned; ++f) {
-    if (face_cell_ids_[f].size() == 1) { extface_ids.push_back(face_map->GID(f)); }
-  }
-  maps_owned_[BOUNDARY_FACE] =
-    Teuchos::rcp(new Epetra_Map(-1, extface_ids.size(), &extface_ids[0], 0, *comm_));
-  exterior_face_importer_ = Teuchos::rcp(new Epetra_Import(*maps_owned_[BOUNDARY_FACE], *face_map));
-
-  // ghosted maps: use the bg mesh to communicate the new GIDs into their ghost values
-  // CELL:
-  if (bg_mesh_->get_comm()->NumProc() == 1) {
-    maps_used_[CELL] = maps_owned_[CELL];
-    maps_used_[FACE] = maps_owned_[FACE];
-  } else {
-    int ncells_bg_owned = bg_mesh_->num_entities(CELL, Parallel_type::OWNED);
-    int ncells_bg_used = bg_mesh_->num_entities(CELL, Parallel_type::ALL);
-    int ncells_log = log_mesh_->num_entities(CELL, Parallel_type::OWNED);
-    int ncells_my_used = num_entities(CELL, Parallel_type::ALL);
-
-    // -- create a populate the owned GIDs
-    Epetra_IntVector bg_gids_owned_c(bg_mesh_->cell_map(false));
-
-    Epetra_Map& my_cell_map = *maps_owned_[CELL];
-    for (int c = 0; c != ncells_bg_owned; ++c) {
-      bg_gids_owned_c[c] = my_cell_map.GID(c + ncells_log);
-    }
-
-    // -- create the map from owned to used
-    Epetra_Import cell_import(bg_mesh_->cell_map(true), bg_mesh_->cell_map(false));
-
-    // -- create the used GIDs vector, comm_unicate
-    Epetra_IntVector bg_gids_used_c(bg_mesh_->cell_map(true));
-    bg_gids_used_c.Import(bg_gids_owned_c, cell_import, Insert);
-
-    // -- create a GID list of the used components
-    Entity_ID_List cells_my_used(ncells_my_used);
-    for (int c = 0; c != ncells_log; ++c) { cells_my_used[c] = my_cell_map.GID(c); }
-    for (int c = 0; c != ncells_bg_used; ++c) { cells_my_used[c + ncells_log] = bg_gids_used_c[c]; }
-
-    // -- create the map
-    maps_used_[CELL] =
-      Teuchos::rcp(new Epetra_Map(-1, ncells_my_used, &cells_my_used[0], 0, *comm_));
-
-    // FACE:
-    int nfaces_bg_owned = bg_mesh_->num_entities(FACE, Parallel_type::OWNED);
-    int nfaces_bg_used = bg_mesh_->num_entities(FACE, Parallel_type::ALL);
-    int nfaces_my_used = face_cell_ids_.size();
-    int nfaces_log_extra = nfaces_my_used - nfaces_bg_used;
-
-    // -- create a populate the owned GIDs
-    Epetra_IntVector bg_gids_owned_f(bg_mesh_->face_map(false));
-
-    Epetra_Map& my_face_map = *maps_owned_[FACE];
-    for (int f = 0; f != nfaces_bg_owned; ++f) {
-      bg_gids_owned_f[f] = my_face_map.GID(f + nfaces_log_extra);
-    }
-
-    // -- create the map from owned to used
-    Epetra_Import face_import(bg_mesh_->face_map(true), bg_mesh_->face_map(false));
-
-    // -- create the used GIDs vector, comm_unicate
-    Epetra_IntVector bg_gids_used_f(bg_mesh_->face_map(true));
-    bg_gids_used_f.Import(bg_gids_owned_f, face_import, Insert);
-
-    // -- create a GID list of the used components
-    Entity_ID_List faces_my_used(nfaces_my_used);
-    for (int f = 0; f != nfaces_log_extra; ++f) { faces_my_used[f] = my_face_map.GID(f); }
-    for (int f = 0; f != nfaces_bg_used; ++f) {
-      faces_my_used[f + nfaces_log_extra] = bg_gids_used_f[f];
-    }
-
-    // -- create the map
-    maps_used_[FACE] =
-      Teuchos::rcp(new Epetra_Map(-1, nfaces_my_used, &faces_my_used[0], 0, *comm_));
-  }
-}
-
-
-// Get parallel type of entity - OWNED, GHOST, ALL (See MeshDefs.hh)
-Parallel_type
-MeshEmbeddedLogical::entity_get_ptype(const Entity_kind kind, const Entity_ID entid) const
-{
-  return entid < num_entities_owned_.at(kind) ? Parallel_type::OWNED : Parallel_type::ALL;
-}
 
 // Parent entity in the source mesh if mesh was derived from another mesh
 Entity_ID
-MeshEmbeddedLogical::entity_get_parent(const Entity_kind kind, const Entity_ID entid) const
+MeshEmbeddedLogical::getEntityParent(const Entity_kind kind, const Entity_ID entid) const
 {
-  int nents_log = log_mesh_->num_entities(kind, Parallel_type::OWNED);
-  return entid < nents_log ? entid : entid - nents_log;
+  if (kind == Entity_kind::CELL) {
+    auto ncells_log = log_mesh_->getNumEntities(kind, Parallel_type::OWNED);
+    if (entid < ncells_log) return entid;
+    else return entid - ncells_log;
+  } else if (kind == Entity_kind::FACE) {
+    auto nfaces_log = log_mesh_->getNumEntities(kind, Parallel_type::OWNED);
+    int nfaces_extra = extra_face_cell_ids_.size<MemSpace_type::HOST>();
+    if (entid < nfaces_log) return entid;
+    else if (entid >= (nfaces_log + nfaces_extra)) return entid - nfaces_log - nfaces_extra;
+    else return -1;
+  } else {
+    return -1;
+  }
 }
 
-
-// Get cell type - UNKNOWN, TRI, QUAD, POLYGON, TET, PRISM, PYRAMID, HEX, POLYHED
-// See MeshDefs.hh
 Cell_type
-MeshEmbeddedLogical::cell_get_type(const Entity_ID cellid) const
+MeshEmbeddedLogical::getCellType(const Entity_ID cellid) const
 {
-  return cellid < log_mesh_->num_entities(CELL, Parallel_type::OWNED) ?
-           log_mesh_->cell_get_type(entity_get_parent(CELL, cellid)) :
-           bg_mesh_->cell_get_type(entity_get_parent(CELL, cellid));
+  auto ncells_log = log_mesh_->getNumEntities(Entity_kind::CELL, Parallel_type::OWNED);
+  if (cellid < ncells_log) return log_mesh_->getCellType(getEntityParent(Entity_kind::CELL, cellid));
+  else return bg_mesh_->getCellType(getEntityParent(Entity_kind::CELL, cellid));
+}
+
+std::size_t
+MeshEmbeddedLogical::getNumEntities(const Entity_kind kind,
+        const Parallel_type ptype) const
+{
+  auto count = log_mesh_->getNumEntities(kind, ptype)
+    + bg_mesh_->getNumEntities(kind, ptype);
+  if (kind == Entity_kind::FACE) count += extra_face_cell_ids_.size<MemSpace_type::HOST>();
+  return count;
+}
+
+
+Entity_GID
+MeshEmbeddedLogical::getEntityGID(const Entity_kind kind, const Entity_ID lid) const
+{
+  if (getComm()->NumProc() == 1) return lid;
+  if (kind == Entity_kind::CELL) {
+    return cell_map_->GID(lid);
+  } else if (kind == Entity_kind::FACE) {
+    return face_map_->GID(lid);
+  } else {
+    Errors::Message msg;
+    msg << "MeshEmbeddedLogical does not have entities of kind " << to_string(kind);
+    Exceptions::amanzi_throw(msg);
+    return -1;
+  }
 }
 
 
 //
-// General mesh information
-// -------------------------
+// Nodal methods
 //
-// Number of entities of any kind (cell, face, node) and in a
-// particular category (OWNED, GHOST, ALL)
-unsigned int
-MeshEmbeddedLogical::num_entities(const Entity_kind kind, const Parallel_type ptype) const
+AmanziGeometry::Point
+MeshEmbeddedLogical::getNodeCoordinate(const Entity_ID node) const
 {
-  return ptype == Parallel_type::OWNED ? num_entities_owned_.at(kind) : num_entities_used_.at(kind);
+  Errors::Message mesg("There are no nodes in a MeshEmbeddedLogical.");
+  Exceptions::amanzi_throw(mesg);
+  return AmanziGeometry::Point();
 }
 
-// Global ID of any entity
-Entity_ID
-MeshEmbeddedLogical::GID(const Entity_ID lid, const Entity_kind kind) const
-{
-  return maps_used_.at(kind)->GID(lid);
-}
-
-
-//
-// Mesh Entity Adjacencies
-//-------------------------
-// Downward Adjacencies
-//---------------------
-// Get nodes of a cell
 void
-MeshEmbeddedLogical::cell_get_nodes(const Entity_ID cellid, Entity_ID_List* nodeids) const
+MeshEmbeddedLogical::getFaceNodes(const Entity_ID f, Entity_ID_List& nodes) const
 {
-  Errors::Message mesg("No nodes in MeshEmbeddedLogical.");
+  Errors::Message mesg("There are no nodes in a MeshEmbeddedLogical.");
   Exceptions::amanzi_throw(mesg);
 }
 
+void
+MeshEmbeddedLogical::getNodeFaces(const Entity_ID nodeid,
+                  const Parallel_type ptype,
+                  Entity_ID_List& faceids) const
+{
+  Errors::Message mesg("There are no nodes in a MeshEmbeddedLogical.");
+  Exceptions::amanzi_throw(mesg);
+}
+
+
+//
+// These are the important ones -- MeshLogical defines cell quantities
+//
+void
+MeshEmbeddedLogical::getCellFacesAndDirs(
+  const Entity_ID c,
+  Entity_ID_List& faces,
+  Entity_Direction_List * const dirs) const
+{
+  auto ncells_log = log_mesh_->getNumEntities(Entity_kind::CELL, Parallel_type::OWNED);
+  if (c < ncells_log) {
+    log_mesh_->getCellFacesAndDirs(c, faces, dirs);
+
+    // check for extras
+    auto nfaces_log = log_mesh_->getNumEntities(Entity_kind::FACE, Parallel_type::OWNED);
+    for (int f=0; f!=extra_face_cell_ids_.size<MemSpace_type::HOST>(); ++f) {
+      if (c == extra_face_cell_ids_.get<MemSpace_type::HOST>(f,0)) {
+        faces.push_back(f + nfaces_log);
+        if (dirs) dirs->push_back(1); // direction always from log to bg
+      }
+    }
+
+  } else {
+    auto nfaces_log = log_mesh_->getNumEntities(Entity_kind::FACE, Parallel_type::OWNED);
+    auto nfaces_extra = extra_face_cell_ids_.size<MemSpace_type::HOST>();
+    bg_mesh_->getCellFacesAndDirs(c-ncells_log, faces, dirs);
+    for (int i=0; i!=faces.size(); ++i) faces[i] += (nfaces_log + nfaces_extra);
+
+    // check for extras
+    auto nfaces_bg = bg_mesh_->getNumEntities(Entity_kind::FACE, Parallel_type::OWNED);
+    for (int f=0; f!=extra_face_cell_ids_.size<MemSpace_type::HOST>(); ++f) {
+      if (c == (extra_face_cell_ids_.get<MemSpace_type::HOST>(f,1)+ncells_log)) {
+        faces.push_back(f + nfaces_log);
+        if (dirs) dirs->push_back(-1); // direction always from log to bg
+      }
+    }
+  }
+}
 
 // Get the bisectors, i.e. vectors from cell centroid to face centroids.
 void
-MeshEmbeddedLogical::cell_get_faces_and_bisectors(const Entity_ID cellid,
-                                                  Entity_ID_List* faceids,
-                                                  std::vector<AmanziGeometry::Point>* bisectors,
-                                                  const bool ordered) const
+MeshEmbeddedLogical::getCellFacesAndBisectors(
+  const Entity_ID c,
+  Entity_ID_List& faces,
+  Point_List * const bisectors) const
 {
-  if (faceids) *faceids = cell_face_ids_[cellid];
-  if (bisectors) *bisectors = cell_face_bisectors_[cellid];
-}
+  auto ncells_log = log_mesh_->getNumEntities(Entity_kind::CELL, Parallel_type::OWNED);
+  if (c < ncells_log) {
+    log_mesh_->getCellFacesAndBisectors(c, faces, bisectors);
 
+    // check for extras
+    auto nfaces_log = log_mesh_->getNumEntities(Entity_kind::FACE, Parallel_type::OWNED);
+    for (int f=0; f!=extra_face_cell_ids_.size<MemSpace_type::HOST>(); ++f) {
+      if (c == extra_face_cell_ids_.get<MemSpace_type::HOST>(f,0)) {
+        faces.push_back(f + nfaces_log);
+        if (bisectors) bisectors->push_back(extra_face_cell_bisectors_.get<MemSpace_type::HOST>(f,0));
+      }
+    }
 
-// Get nodes of face
-// On a distributed mesh, all nodes ( Parallel_type::OWNED or Parallel_type::GHOST) of the face
-// are returned
-// In 3D, the nodes of the face are returned in ccw order consistent
-// with the face normal
-// In 2D, nfnodes is 2
-void
-MeshEmbeddedLogical::face_get_nodes(const Entity_ID faceid, Entity_ID_List* nodeids) const
-{
-  Errors::Message mesg("No nodes in MeshEmbeddedLogical.");
-  Exceptions::amanzi_throw(mesg);
-}
-
-
-// Get nodes of edge
-void
-MeshEmbeddedLogical::edge_get_nodes(const Entity_ID edgeid,
-                                    Entity_ID* nodeid0,
-                                    Entity_ID* nodeid1) const
-{
-  Errors::Message mesg("No nodes in MeshEmbeddedLogical.");
-  Exceptions::amanzi_throw(mesg);
-}
-
-// Upward adjacencies
-//-------------------
-
-// Cells of type 'ptype' connected to a node - The order of cells is
-// not guaranteed to be the same for corresponding nodes on
-// different processors
-void
-MeshEmbeddedLogical::node_get_cells(const Entity_ID nodeid,
-                                    const Parallel_type ptype,
-                                    Entity_ID_List* cellids) const
-{
-  Errors::Message mesg("No nodes in MeshEmbeddedLogical.");
-  Exceptions::amanzi_throw(mesg);
-}
-
-// Faces of type 'ptype' connected to a node - The order of faces is
-// not guarnateed to be the same for corresponding nodes on
-// different processors
-void
-MeshEmbeddedLogical::node_get_faces(const Entity_ID nodeid,
-                                    const Parallel_type ptype,
-                                    Entity_ID_List* faceids) const
-{
-  Errors::Message mesg("No nodes in MeshEmbeddedLogical.");
-  Exceptions::amanzi_throw(mesg);
-}
-
-
-// Same level adjacencies
-//-----------------------
-
-// Face connected neighboring cells of given cell of a particular ptype
-// (e.g. a hex has 6 face neighbors)
-
-// The order in which the cellids are returned cannot be
-// guaranteed in general except when ptype = ALL, in which case
-// the cellids will correcpond to cells across the respective
-// faces given by cell_get_faces
-void
-MeshEmbeddedLogical::cell_get_face_adj_cells(const Entity_ID cellid,
-                                             const Parallel_type ptype,
-                                             Entity_ID_List* fadj_cellids) const
-{
-  Errors::Message mesg(
-    "Not implemented."); // this could be implemented, just not sure we need it yet. --etc
-  Exceptions::amanzi_throw(mesg);
-}
-
-
-//
-// Mesh entity geometry
-//--------------
-//
-// Node coordinates - 3 in 3D and 2 in 2D
-void
-MeshEmbeddedLogical::node_get_coordinates(const Entity_ID nodeid,
-                                          AmanziGeometry::Point* ncoord) const
-{
-  Errors::Message mesg("No nodes in MeshEmbeddedLogical.");
-  Exceptions::amanzi_throw(mesg);
-}
-
-
-//
-// Mesh modification
-//-------------------
-
-// Set coordinates of node
-void
-MeshEmbeddedLogical::node_set_coordinates(const Entity_ID nodeid,
-                                          const AmanziGeometry::Point ncoord)
-{
-  Errors::Message mesg("No nodes in MeshEmbeddedLogical.");
-  Exceptions::amanzi_throw(mesg);
-}
-
-void
-MeshEmbeddedLogical::node_set_coordinates(const Entity_ID nodeid, const double* ncoord)
-{
-  Errors::Message mesg("No nodes in MeshEmbeddedLogical.");
-  Exceptions::amanzi_throw(mesg);
-}
-
-
-// Deformation not supported.
-int
-MeshEmbeddedLogical::deform(const std::vector<double>& target_cell_volumes_in,
-                            const std::vector<double>& min_cell_volumes_in,
-                            const Entity_ID_List& fixed_nodes,
-                            const bool move_vertical)
-{
-  Errors::Message mesg("No nodes in MeshEmbeddedLogical.");
-  Exceptions::amanzi_throw(mesg);
-  return -1;
-}
-
-//
-// Epetra maps
-//------------
-const Epetra_Map&
-MeshEmbeddedLogical::cell_map(bool include_ghost) const
-{
-  return include_ghost ? *maps_used_.at(CELL) : *maps_owned_.at(CELL);
-}
-
-const Epetra_Map&
-MeshEmbeddedLogical::face_map(bool include_ghost) const
-{
-  return include_ghost ? *maps_used_.at(FACE) : *maps_owned_.at(FACE);
-}
-
-const Epetra_Map&
-MeshEmbeddedLogical::node_map(bool include_ghost) const
-{
-  Errors::Message mesg("No nodes in MeshEmbeddedLogical.");
-  Exceptions::amanzi_throw(mesg);
-  return include_ghost ? *maps_used_.at(NODE) : *maps_owned_.at(NODE);
-}
-
-
-const Epetra_Map&
-MeshEmbeddedLogical::exterior_face_map(bool include_ghost) const
-{
-  return *maps_owned_.at(BOUNDARY_FACE);
-}
-
-// Epetra importer that will allow apps to import values from a
-// Epetra vector defined on all owned faces into an Epetra vector
-// defined only on exterior faces
-const Epetra_Import&
-MeshEmbeddedLogical::exterior_face_importer(void) const
-{
-  return *exterior_face_importer_;
-}
-
-
-const Epetra_Map&
-MeshEmbeddedLogical::exterior_node_map(bool include_ghost) const
-{
-  Errors::Message mesg("No exterior node map in MeshEmbeddedLogical.");
-  Exceptions::amanzi_throw(mesg);
-  throw(mesg);
-}
-
-
-//
-// Mesh Sets for ICs, BCs, Material Properties and whatever else
-//--------------------------------------------------------------
-
-// Get list of entities of type 'category' in set
-// FIX ME: This is probably not generic enough. --etc
-// OBSOLETE: use region name bassed routine.
-void
-MeshEmbeddedLogical::get_set_entities(const std::string& setname,
-                                      const Entity_kind kind,
-                                      const Parallel_type ptype,
-                                      Entity_ID_List* entids) const
-{
-  Teuchos::RCP<const AmanziGeometry::Region> rgn = geometric_model_->FindRegion(setname);
-
-  if (rgn->get_type() == AmanziGeometry::RegionType::ALL) {
-    int nent = num_entities(kind, ptype);
-    entids->resize(num_entities(kind, ptype));
-    for (int i = 0; i != nent; ++i) { (*entids)[i] = i; }
-    return;
-  }
-
-  // ASSUMES that logical mesh is EnumeratedSets, bg mesh is all others.
-  // This is bad.  --etc
-  if (rgn->get_type() == AmanziGeometry::RegionType::ENUMERATED) {
-    log_mesh_->get_set_entities(rgn->get_name(), kind, ptype, entids);
   } else {
-    bg_mesh_->get_set_entities(rgn->get_name(), kind, ptype, entids);
+    auto nfaces_log = log_mesh_->getNumEntities(Entity_kind::FACE, Parallel_type::OWNED);
+    auto nfaces_extra = extra_face_cell_ids_.size<MemSpace_type::HOST>();
+    bg_mesh_->getCellFacesAndBisectors(c-ncells_log, faces, bisectors);
+    for (int i=0; i!=faces.size(); ++i) faces[i] += (nfaces_log + nfaces_extra);
 
-    // shift local indices
-    Entity_ID shift = num_entities(kind, ptype) - bg_mesh_->num_entities(kind, ptype);
-    for (Entity_ID_List::iterator i = entids->begin(); i != entids->end(); ++i) { (*i) += shift; }
-  }
-  return;
-}
-
-
-void
-MeshEmbeddedLogical::get_set_entities_and_vofs(const std::string& setname,
-                                               const Entity_kind kind,
-                                               const Parallel_type ptype,
-                                               Entity_ID_List* entids,
-                                               std::vector<double>* vofs) const
-{
-  get_set_entities(setname, kind, ptype, entids);
-  return;
-}
-
-
-// Miscellaneous functions
-void
-MeshEmbeddedLogical::write_to_exodus_file(const std::string filename) const
-{
-  // don't know how to do this! FIXME! --etc
-}
-
-
-// Geometry
-int
-MeshEmbeddedLogical::compute_cell_geometry_(const Entity_ID cellid,
-                                            double* volume,
-                                            AmanziGeometry::Point* centroid) const
-{
-  // this is a placeholder, these cannot be recomputed
-  if (volume) *volume = cell_volumes_[cellid];
-
-  if (centroid) {
-    if (cell_centroids_.size() > 0) {
-      *centroid = cell_centroids_[cellid];
-    } else {
-      *centroid = AmanziGeometry::Point();
+    // check for extras
+    auto nfaces_bg = bg_mesh_->getNumEntities(Entity_kind::FACE, Parallel_type::OWNED);
+    for (int f=0; f!=extra_face_cell_ids_.size<MemSpace_type::HOST>(); ++f) {
+      if (c == (extra_face_cell_ids_.get<MemSpace_type::HOST>(f,1)+ncells_log)) {
+        faces.push_back(f + nfaces_log);
+        if (bisectors) bisectors->push_back(extra_face_cell_bisectors_.get<MemSpace_type::HOST>(f,1));
+      }
     }
   }
-  return 1;
 }
 
-
-int
-MeshEmbeddedLogical::compute_face_geometry_(const Entity_ID faceid,
-                                            double* area,
-                                            AmanziGeometry::Point* centroid,
-                                            std::vector<AmanziGeometry::Point>* normals) const
+double
+MeshEmbeddedLogical::getCellVolume(const Entity_ID c) const
 {
-  // this is a placeholder, these cannot be recomputed
-  if (area) *area = face_areas_[faceid];
-  if (centroid) *centroid = AmanziGeometry::Point();
-  *normals = face_normals_[faceid];
-  return 1;
+  auto ncells_log = log_mesh_->getNumEntities(Entity_kind::CELL, Parallel_type::OWNED);
+  if (c < ncells_log) return log_mesh_->getCellVolume(c);
+  else return bg_mesh_->getCellVolume(c - ncells_log);
+}
+
+AmanziGeometry::Point
+MeshEmbeddedLogical::getCellCentroid(const Entity_ID c) const
+{
+  auto ncells_log = log_mesh_->getNumEntities(Entity_kind::CELL, Parallel_type::OWNED);
+  if (c < ncells_log) return log_mesh_->getCellCentroid(c);
+  else return bg_mesh_->getCellCentroid(c - ncells_log);
 }
 
 
-// get faces of a cell and directions in which it is used - this function
-// is implemented in each mesh framework. The results are cached in
-// the base class
 void
-MeshEmbeddedLogical::cell_get_faces_and_dirs_internal_(const Entity_ID cellid,
-                                                       Entity_ID_List* faceids,
-                                                       std::vector<int>* face_dirs,
-                                                       const bool ordered) const
+MeshEmbeddedLogical::getFaceCells(const Entity_ID f,
+        const Parallel_type ptype,
+        Entity_ID_List& cells) const
 {
-  Errors::Message mesg("DEVELOPER ERROR: cell_get_faces_and_dirs_internal_() should not be called");
-  Exceptions::amanzi_throw(mesg);
+  auto ncells_log = log_mesh_->getNumEntities(Entity_kind::CELL, Parallel_type::OWNED);
+  auto nfaces_log = log_mesh_->getNumEntities(Entity_kind::FACE, Parallel_type::OWNED);
+  auto nfaces_extra = extra_face_cell_ids_.size<MemSpace_type::HOST>();
+  if (f < nfaces_log) {
+    log_mesh_->getFaceCells(f, ptype, cells);
+  } else if (f >= nfaces_log + nfaces_extra) {
+    bg_mesh_->getFaceCells(f - nfaces_log - nfaces_extra, ptype, cells);
+    for (auto& c : cells) c += ncells_log;
+  } else {
+    cells = asVector(extra_face_cell_ids_.getRow<MemSpace_type::HOST>(f - nfaces_log));
+    cells[1] += ncells_log;
+  }
 }
 
-// Cells connected to a face - this function is implemented in each
-// mesh framework. The results are cached in the base class
-void
-MeshEmbeddedLogical::face_get_cells_internal_(const Entity_ID faceid,
-                                              const Parallel_type ptype,
-                                              Entity_ID_List* cellids) const
+double
+MeshEmbeddedLogical::getFaceArea(const Entity_ID f) const
 {
-  Errors::Message mesg("DEVELOPER ERROR: face_get_cells_internal_() should not be called");
-  Exceptions::amanzi_throw(mesg);
+  auto nfaces_log = log_mesh_->getNumEntities(Entity_kind::FACE, Parallel_type::OWNED);
+  auto nfaces_extra = extra_face_cell_ids_.size<MemSpace_type::HOST>();
+  if (f < nfaces_log) {
+    return log_mesh_->getFaceArea(f);
+  } else if (f >= nfaces_log + nfaces_extra) {
+    return bg_mesh_->getFaceArea(f - nfaces_log - nfaces_extra);
+  } else {
+    return AmanziGeometry::norm(extra_face_area_normals_[f - nfaces_log]);
+  }
 }
 
-// edges of a face - this function is implemented in each mesh
-// framework. The results are cached in the base class
-void
-MeshEmbeddedLogical::face_get_edges_and_dirs_internal_(const Entity_ID faceid,
-                                                       Entity_ID_List* edgeids,
-                                                       std::vector<int>* edge_dirs,
-                                                       const bool ordered) const
+AmanziGeometry::Point
+MeshEmbeddedLogical::getFaceCentroid(const Entity_ID f) const
 {
-  Errors::Message mesg("No edges in MeshEmbeddedLogical.");
-  Exceptions::amanzi_throw(mesg);
+  auto nfaces_log = log_mesh_->getNumEntities(Entity_kind::FACE, Parallel_type::OWNED);
+  auto nfaces_extra = extra_face_cell_ids_.size<MemSpace_type::HOST>();
+  if (f < nfaces_log) {
+    return log_mesh_->getFaceCentroid(f);
+  } else if (f >= nfaces_log + nfaces_extra) {
+    return bg_mesh_->getFaceCentroid(f - nfaces_log - nfaces_extra);
+  } else {
+    auto log_cc = log_mesh_->getCellCentroid(extra_face_cell_ids_.get<MemSpace_type::HOST>(f - nfaces_log,0));
+    auto bg_cc = bg_mesh_->getCellCentroid(extra_face_cell_ids_.get<MemSpace_type::HOST>(f - nfaces_log,1));
+    return (log_cc + bg_cc) / 2.;
+  }
+}
+
+AmanziGeometry::Point
+MeshEmbeddedLogical::getFaceNormal(const Entity_ID f,
+        const Entity_ID c, int * const orientation) const
+{
+  auto ncells_log = log_mesh_->getNumEntities(Entity_kind::CELL, Parallel_type::OWNED);
+  auto nfaces_log = log_mesh_->getNumEntities(Entity_kind::FACE, Parallel_type::OWNED);
+  auto nfaces_extra = extra_face_cell_ids_.size<MemSpace_type::HOST>();
+
+  if (f < nfaces_log) {
+    return log_mesh_->getFaceNormal(f, c, orientation);
+  } else if (f >= nfaces_log + nfaces_extra) {
+    return bg_mesh_->getFaceNormal(f - nfaces_log - nfaces_extra, c - ncells_log, orientation);
+  } else {
+    AmanziGeometry::Point normal;
+    if (c < ncells_log) {
+      normal = extra_face_area_normals_[f - nfaces_log];
+      if (orientation) *orientation = 1;
+    } else {
+      normal = -extra_face_area_normals_[f - nfaces_log];
+      if (orientation) *orientation = -1;
+    }
+    return normal;
+  }
 }
 
 
-// edges of a cell - this function is implemented in each mesh
-// framework. The results are cached in the base class.
-void
-MeshEmbeddedLogical::cell_get_edges_internal_(const Entity_ID cellid, Entity_ID_List* edgeids) const
-{
-  Errors::Message mesg("No edges in MeshEmbeddedLogical.");
-  Exceptions::amanzi_throw(mesg);
-}
 
 
-int
-MeshEmbeddedLogical::build_columns_() const
-{
-  Errors::Message mesg("No columns are buildable in MeshEmbeddedLogical.");
-  Exceptions::amanzi_throw(mesg);
-  return -1;
-}
 
-} // namespace AmanziMesh
-} // namespace Amanzi
+
+
+
+
+
+
+} // close namespace AmanziMesh
+} // close namespace Amanzi
