@@ -24,18 +24,15 @@ namespace Amanzi {
 namespace AmanziMesh {
 
 
-MeshLogicalAudit::MeshLogicalAudit(const Teuchos::RCP<const AmanziMesh::Mesh>& mesh_,
-                                   std::ostream& os_)
-  : mesh(mesh_),
-    comm_(mesh_->get_comm()),
-    MyPID(mesh_->get_comm()->MyPID()),
-    nface(mesh_->face_map(true).NumMyElements()),
-    ncell(mesh_->cell_map(true).NumMyElements()),
+MeshLogicalAudit:: MeshLogicalAudit(const Teuchos::RCP<const AmanziMesh::Mesh> &mesh_, std::ostream& os_) :
+    mesh(mesh_),
+    comm_(mesh_->getComm()),
+    MyPID(mesh_->getComm()->MyPID()),
     os(os_),
+    nface(mesh_->getMap(AmanziMesh::Entity_kind::FACE, true).NumMyElements()),
+    ncell(mesh_->getMap(AmanziMesh::Entity_kind::CELL, true).NumMyElements()),
     MAX_OUT(5)
-{
-  create_test_dependencies();
-}
+{ create_test_dependencies(); }
 
 // Verify runs all the tests in an order that respects the dependencies
 // between tests.  If a particular test fails, all other tests that have it
@@ -43,8 +40,7 @@ MeshLogicalAudit::MeshLogicalAudit(const Teuchos::RCP<const AmanziMesh::Mesh>& m
 // a collective fail/pass result in parallel, so that all processes proceed
 // through the tests in lockstep.
 
-int
-MeshLogicalAudit::Verify() const
+int MeshLogicalAudit::Verify() const
 {
   int status = 0;
 
@@ -81,8 +77,7 @@ MeshLogicalAudit::Verify() const
 // The last line specifies that other_test_handle is a pre-requisite for
 // my_test_handle.  There may be multiple pre-requisites or none.
 
-void
-MeshLogicalAudit::create_test_dependencies()
+void MeshLogicalAudit::create_test_dependencies()
 {
   // Entity_counts tests
   Graph::vertex_descriptor test01 = add_vertex(g);
@@ -109,7 +104,7 @@ MeshLogicalAudit::create_test_dependencies()
   g[test08].name = "cell_to_face_consistency";
   g[test08].test = &MeshLogicalAudit::check_faces_cell_consistency;
   add_edge(test05, test08, g);
-
+  
   // cell degeneracy test
   Graph::vertex_descriptor test10 = add_vertex(g);
   g[test10].name = "topological non-degeneracy of cells";
@@ -131,7 +126,7 @@ MeshLogicalAudit::create_test_dependencies()
   g[test14].name = "face-cell-bisector geometry";
   g[test14].test = &MeshLogicalAudit::check_cell_face_bisector_geometry;
   add_edge(test08, test14, g);
-
+  
   Graph::vertex_descriptor test17 = add_vertex(g);
   g[test17].name = "owned and overlap face maps";
   g[test17].test = &MeshLogicalAudit::check_face_maps;
@@ -153,6 +148,7 @@ MeshLogicalAudit::create_test_dependencies()
   g[test32].test = &MeshLogicalAudit::check_face_partition;
   add_edge(test17, test32, g);
   add_edge(test18, test32, g);
+  
 }
 
 ////////////////////////////////////////////////////////////////////////////////
@@ -173,39 +169,38 @@ MeshLogicalAudit::create_test_dependencies()
 // if any discrepancy is found, but it is safe to perform other tests as
 // they do not use these methods.
 
-bool
-MeshLogicalAudit::check_entity_counts() const
+bool MeshLogicalAudit::check_entity_counts() const
 {
   int n, nref;
   bool error = false;
 
   // Check the number of owned faces.
-  n = mesh->num_entities(AmanziMesh::FACE, AmanziMesh::Parallel_type::OWNED);
-  nref = mesh->face_map(false).NumMyElements();
+  n = mesh->getNumEntities(AmanziMesh::Entity_kind::FACE,AmanziMesh::Parallel_type::OWNED);
+  nref = mesh->getMap(AmanziMesh::Entity_kind::FACE, false).NumMyElements();
   if (n != nref) {
     os << "ERROR: num_entities(FACE,OWNED)=" << n << "; should be " << nref << std::endl;
     error = true;
   }
 
   // Check the number of used faces.
-  n = mesh->num_entities(AmanziMesh::FACE, AmanziMesh::Parallel_type::ALL);
-  nref = mesh->face_map(true).NumMyElements();
+  n = mesh->getNumEntities(AmanziMesh::Entity_kind::FACE,AmanziMesh::Parallel_type::ALL);
+  nref = mesh->getMap(AmanziMesh::Entity_kind::FACE, true).NumMyElements();
   if (n != nref) {
     os << "ERROR: num_entities(FACE,ALL)=" << n << "; should be " << nref << std::endl;
     error = true;
   }
 
   // Check the number of owned cells.
-  n = mesh->num_entities(AmanziMesh::CELL, AmanziMesh::Parallel_type::OWNED);
-  nref = mesh->cell_map(false).NumMyElements();
+  n = mesh->getNumEntities(AmanziMesh::Entity_kind::CELL,AmanziMesh::Parallel_type::OWNED);
+  nref = mesh->getMap(AmanziMesh::Entity_kind::CELL, false).NumMyElements();
   if (n != nref) {
     os << "ERROR: num_entities(CELL,OWNED)=" << n << "; should be " << nref << std::endl;
     error = true;
   }
 
   // Check the number of used cells.
-  n = mesh->num_entities(AmanziMesh::CELL, AmanziMesh::Parallel_type::ALL);
-  nref = mesh->cell_map(true).NumMyElements();
+  n = mesh->getNumEntities(AmanziMesh::Entity_kind::CELL,AmanziMesh::Parallel_type::ALL);
+  nref = mesh->getMap(AmanziMesh::Entity_kind::CELL, true).NumMyElements();
   if (n != nref) {
     os << "ERROR: num_entities(CELL,ALL)=" << n << "; should be " << nref << std::endl;
     error = true;
@@ -215,12 +210,11 @@ MeshLogicalAudit::check_entity_counts() const
 }
 // Returns true if the values in the list are distinct -- no repeats.
 
-bool
-MeshLogicalAudit::distinct_values(const AmanziMesh::Entity_ID_List& list) const
+bool MeshLogicalAudit::distinct_values(const AmanziMesh::Entity_ID_List &list) const
 {
   AmanziMesh::Entity_ID_List copy(list);
   sort(copy.begin(), copy.end());
-  return (adjacent_find(copy.begin(), copy.end()) == copy.end());
+  return (adjacent_find(copy.begin(),copy.end()) == copy.end());
 }
 
 
@@ -229,8 +223,7 @@ MeshLogicalAudit::distinct_values(const AmanziMesh::Entity_ID_List& list) const
 // the alternative accessors is checked elsewhere.  A nonzero return value
 // signals an error, and further tests using its data should be avoided.
 
-bool
-MeshLogicalAudit::check_cell_to_faces() const
+bool MeshLogicalAudit::check_cell_to_faces() const
 {
   AmanziMesh::Entity_ID_List bad_cells, bad_cells1;
   AmanziMesh::Entity_ID_List bad_faces;
@@ -239,7 +232,7 @@ MeshLogicalAudit::check_cell_to_faces() const
 
   for (AmanziMesh::Entity_ID j = 0; j < ncell; ++j) {
     try {
-      mesh->cell_get_faces(j, &cface); // this may fail
+      mesh->getCellFaces(j, cface); // this may fail
       bool invalid_refs = false;
       for (int k = 0; k < cface.size(); ++k) {
         if (cface[k] >= nface) invalid_refs = true;
@@ -273,14 +266,13 @@ MeshLogicalAudit::check_cell_to_faces() const
 // nonzero return value indicates that faces were found that do not belong
 // to any cell or belong to more than two cells (bad topology).
 
-bool
-MeshLogicalAudit::check_face_refs_by_cells() const
+bool MeshLogicalAudit::check_face_refs_by_cells() const
 {
   AmanziMesh::Entity_ID_List cface;
   vector<unsigned int> refs(nface, 0);
 
   for (AmanziMesh::Entity_ID j = 0; j < ncell; ++j) {
-    mesh->cell_get_faces(j, &cface);
+    mesh->getCellFaces(j, cface);
     for (int k = 0; k < cface.size(); ++k) (refs[cface[k]])++;
   }
 
@@ -313,27 +305,30 @@ MeshLogicalAudit::check_face_refs_by_cells() const
 
 
 // Check the consistency of face_get_cells and cell_get_faces.
-bool
-MeshLogicalAudit::check_faces_cell_consistency() const
+bool MeshLogicalAudit::check_faces_cell_consistency() const
 {
   AmanziMesh::Entity_ID_List cface;
   AmanziMesh::Entity_ID_List fcell;
 
   AmanziMesh::Entity_ID_List bad_cells;
   for (AmanziMesh::Entity_ID j = 0; j < ncell; ++j) {
-    mesh->cell_get_faces(j, &cface);
+    mesh->getCellFaces(j, cface);
     for (int k = 0; k < cface.size(); ++k) {
-      mesh->face_get_cells(cface[k], Parallel_type::ALL, &fcell);
-      if (std::find(fcell.begin(), fcell.end(), j) == fcell.end()) { bad_cells.push_back(j); }
+      mesh->getFaceCells(cface[k], Parallel_type::ALL, fcell);
+      if (std::find(fcell.begin(),fcell.end(), j) == fcell.end()) {
+        bad_cells.push_back(j);
+      }
     }
   }
 
   AmanziMesh::Entity_ID_List bad_faces;
   for (AmanziMesh::Entity_ID j = 0; j < nface; ++j) {
-    mesh->face_get_cells(j, Parallel_type::ALL, &fcell);
+    mesh->getFaceCells(j, Parallel_type::ALL, fcell);
     for (int k = 0; k < fcell.size(); ++k) {
-      mesh->cell_get_faces(fcell[k], &cface);
-      if (std::find(cface.begin(), cface.end(), j) == cface.end()) { bad_faces.push_back(j); }
+      mesh->getCellFaces(fcell[k], cface);
+      if (std::find(cface.begin(),cface.end(), j) == cface.end()) {
+        bad_faces.push_back(j);
+      }
     }
   }
 
@@ -358,8 +353,7 @@ MeshLogicalAudit::check_faces_cell_consistency() const
 // the consistency of the alternative methods is checked elsewhere.  If this
 // test fails, further tests using this data should be avoided.
 
-bool
-MeshLogicalAudit::check_cell_to_face_dirs() const
+bool MeshLogicalAudit::check_cell_to_face_dirs() const
 {
   AmanziMesh::Entity_ID_List faces;
   vector<int> fdirs;
@@ -368,7 +362,7 @@ MeshLogicalAudit::check_cell_to_face_dirs() const
   for (AmanziMesh::Entity_ID j = 0; j < ncell; ++j) {
     fdirs.assign(6, INT_MAX);
     try {
-      mesh->cell_get_faces_and_dirs(j, &faces, &fdirs); // this may fail
+      mesh->getCellFacesAndDirs(j, faces, &fdirs);  // this may fail
       bool bad_data = false;
       for (int k = 0; k < fdirs.size(); ++k)
         if (fdirs[k] != -1 && fdirs[k] != 1) bad_data = true;
@@ -399,8 +393,7 @@ MeshLogicalAudit::check_cell_to_face_dirs() const
 // If this test fails, many further tests involving the cell_to_nodes data
 // should be avoided.
 
-bool
-MeshLogicalAudit::check_cell_degeneracy() const
+bool MeshLogicalAudit::check_cell_degeneracy() const
 {
   os << "Checking cells for topological degeneracy ..." << std::endl;
 
@@ -408,7 +401,7 @@ MeshLogicalAudit::check_cell_degeneracy() const
   AmanziMesh::Entity_ID_List bad_cells;
 
   for (AmanziMesh::Entity_ID j = 0; j < ncell; ++j) {
-    mesh->cell_get_faces(j, &cface); // should not fail
+    mesh->getCellFaces(j, cface); // should not fail
     if (!distinct_values(cface)) bad_cells.push_back(j);
   }
 
@@ -420,12 +413,11 @@ MeshLogicalAudit::check_cell_degeneracy() const
     error = true;
   }
 
-  return global_any(error);
+    return global_any(error);
 }
 
 // The cells must not be degenerate, with zero volume
-bool
-MeshLogicalAudit::check_cell_geometry() const
+bool MeshLogicalAudit::check_cell_geometry() const
 {
   os << "Checking cell geometry ..." << std::endl;
   AmanziGeometry::Point centroid;
@@ -433,7 +425,7 @@ MeshLogicalAudit::check_cell_geometry() const
   AmanziMesh::Entity_ID_List bad_cells;
 
   for (AmanziMesh::Entity_ID j = 0; j < ncell; ++j) {
-    hvol = mesh->cell_volume(j);
+    hvol = mesh->getCellVolume(j);
     if (hvol <= 1.e-10) bad_cells.push_back(j);
   }
 
@@ -450,15 +442,14 @@ MeshLogicalAudit::check_cell_geometry() const
 }
 
 // The faces must not be degenerate
-bool
-MeshLogicalAudit::check_face_geometry() const
+bool MeshLogicalAudit::check_face_geometry() const
 {
   os << "Checking face geometry ..." << std::endl;
   double hvol;
   AmanziMesh::Entity_ID_List bad_faces;
 
   for (AmanziMesh::Entity_ID j = 0; j < nface; ++j) {
-    hvol = mesh->face_area(j);
+    hvol = mesh->getFaceArea(j);
     if (hvol <= 0.0) bad_faces.push_back(j);
   }
 
@@ -475,8 +466,7 @@ MeshLogicalAudit::check_face_geometry() const
 
 
 // The faces must not be degenerate
-bool
-MeshLogicalAudit::check_cell_face_bisector_geometry() const
+bool MeshLogicalAudit::check_cell_face_bisector_geometry() const
 {
   os << "Checking cell-to-face bisector geometry ..." << std::endl;
   AmanziMesh::Entity_ID_List bad_cells;
@@ -484,13 +474,15 @@ MeshLogicalAudit::check_cell_face_bisector_geometry() const
   AmanziMesh::Entity_ID_List cface;
   std::vector<AmanziGeometry::Point> bisectors;
   for (AmanziMesh::Entity_ID j = 0; j < ncell; ++j) {
-    mesh->cell_get_faces_and_bisectors(j, &cface, &bisectors);
+    mesh->getCellFacesAndBisectors(j, cface, &bisectors);
     if (cface.size() != bisectors.size()) {
       bad_cells.push_back(j);
       continue;
     }
-    for (int i = 0; i != cface.size(); ++i) {
-      if (AmanziGeometry::norm(bisectors[i]) < 1.e-10) { bad_cells.push_back(j); }
+    for (int i=0; i!=cface.size(); ++i) {
+      if (AmanziGeometry::norm(bisectors[i]) < 1.e-10) {
+        bad_cells.push_back(j);
+      }
     }
   }
 
@@ -506,20 +498,17 @@ MeshLogicalAudit::check_cell_face_bisector_geometry() const
 }
 
 
-bool
-MeshLogicalAudit::check_face_maps() const
+bool MeshLogicalAudit::check_face_maps() const
 {
-  return check_maps(mesh->face_map(false), mesh->face_map(true));
+  return check_maps(mesh->getMap(AmanziMesh::Entity_kind::FACE, false), mesh->getMap(AmanziMesh::Entity_kind::FACE, true));
 }
 
-bool
-MeshLogicalAudit::check_cell_maps() const
+bool MeshLogicalAudit::check_cell_maps() const
 {
-  return check_maps(mesh->cell_map(false), mesh->cell_map(true));
+  return check_maps(mesh->getMap(AmanziMesh::Entity_kind::CELL, false), mesh->getMap(AmanziMesh::Entity_kind::CELL, true));
 }
 
-bool
-MeshLogicalAudit::check_maps(const Epetra_Map& map_own, const Epetra_Map& map_use) const
+bool MeshLogicalAudit::check_maps(const Epetra_Map &map_own, const Epetra_Map &map_use) const
 {
   bool error = false;
 
@@ -543,14 +532,16 @@ MeshLogicalAudit::check_maps(const Epetra_Map& map_own, const Epetra_Map& map_us
   // necessarily uniformly) across all processers
 
   std::vector<int> owned_GIDs(map_own.NumMyElements());
-  for (int i = 0; i < map_own.NumMyElements(); i++) owned_GIDs[i] = map_own.GID(i);
+  for (int i = 0; i < map_own.NumMyElements(); i++)
+    owned_GIDs[i] = map_own.GID(i);
   std::sort(owned_GIDs.begin(), owned_GIDs.end());
 
-  for (int i = 0; i < map_own.NumMyElements() - 1; i++) {
-    int diff = owned_GIDs[i + 1] - owned_GIDs[i];
+  for (int i = 0; i < map_own.NumMyElements()-1; i++) {
+    int diff = owned_GIDs[i+1]-owned_GIDs[i];
     if (diff > 1) {
       os << "ERROR: owned map is not contiguous" << std::endl;
-      os << "Global IDs jump from " << owned_GIDs[i] << " to " << owned_GIDs[i + 1] << std::endl;
+      os << "Global IDs jump from " << owned_GIDs[i] << " to " <<
+          owned_GIDs[i+1] << std::endl;
       error = true;
     }
   }
@@ -558,7 +549,9 @@ MeshLogicalAudit::check_maps(const Epetra_Map& map_own, const Epetra_Map& map_us
   error = global_any(error);
   if (error) return error;
 
-  if (comm_->NumProc() == 1) {
+  if (comm_->NumProc() == 1)
+  {
+
     // Serial or 1-process MPI
 
     if (!map_use.SameAs(map_own)) {
@@ -568,7 +561,10 @@ MeshLogicalAudit::check_maps(const Epetra_Map& map_own, const Epetra_Map& map_us
 
     return global_any(error);
 
-  } else {
+  }
+  else
+  {
+
     // Multi-process MPI
 
     int num_own = map_own.NumMyElements();
@@ -592,10 +588,10 @@ MeshLogicalAudit::check_maps(const Epetra_Map& map_own, const Epetra_Map& map_us
     if (error) return error;
 
     // Verify that the overlap indices are owned by other processes.
-    int* gids = new int[num_ovl];
-    int* pids = new int[num_ovl];
-    int* lids = new int[num_ovl];
-    for (int j = 0; j < num_ovl; ++j) gids[j] = map_use.GID(j + num_own);
+    int *gids = new int[num_ovl];
+    int *pids = new int[num_ovl];
+    int *lids = new int[num_ovl];
+    for (int j = 0; j < num_ovl; ++j) gids[j] = map_use.GID(j+num_own);
     map_own.RemoteIDList(num_ovl, gids, pids, lids);
     bad_map = false;
     for (int j = 0; j < num_ovl; ++j)
@@ -606,16 +602,16 @@ MeshLogicalAudit::check_maps(const Epetra_Map& map_own, const Epetra_Map& map_us
     }
 
     // Look for duplicates among the overlap indices.
-    vector<int> ovl_gids(gids, gids + num_ovl);
+    vector<int> ovl_gids(gids, gids+num_ovl);
     sort(ovl_gids.begin(), ovl_gids.end());
-    if (adjacent_find(ovl_gids.begin(), ovl_gids.end()) != ovl_gids.end()) {
+    if (adjacent_find(ovl_gids.begin(),ovl_gids.end()) != ovl_gids.end()) {
       os << "ERROR: duplicate ghosts in overlap map." << std::endl;
       error = true;
     }
 
-    delete[] lids;
-    delete[] pids;
-    delete[] gids;
+    delete [] lids;
+    delete [] pids;
+    delete [] gids;
 
     return global_any(error);
   }
@@ -630,12 +626,11 @@ MeshLogicalAudit::check_maps(const Epetra_Map& map_own, const Epetra_Map& map_us
 // the same nodes as the correct face.  So the two faces would be
 // geometrically identical, including orientation, but be distinct.
 
-bool
-MeshLogicalAudit::check_cell_to_faces_ghost_data() const
+bool MeshLogicalAudit::check_cell_to_faces_ghost_data() const
 {
-  const Epetra_Map& face_map = mesh->face_map(true);
-  const Epetra_Map& cell_map_own = mesh->cell_map(false);
-  const Epetra_Map& cell_map_use = mesh->cell_map(true);
+  const Epetra_Map &face_map = mesh->getMap(AmanziMesh::Entity_kind::FACE, true);
+  const Epetra_Map &cell_map_own = mesh->getMap(AmanziMesh::Entity_kind::CELL, false);
+  const Epetra_Map &cell_map_use = mesh->getMap(AmanziMesh::Entity_kind::CELL, true);
 
   int ncell_own = cell_map_own.NumMyElements();
   int ncell_use = cell_map_use.NumMyElements();
@@ -646,16 +641,18 @@ MeshLogicalAudit::check_cell_to_faces_ghost_data() const
   // Create a matrix of the GIDs for all owned cells.
   int maxfaces = 0;
   for (AmanziMesh::Entity_ID j = 0; j < ncell_own; ++j) {
-    mesh->cell_get_faces(j, &cface);
+    mesh->getCellFaces(j, cface);
     maxfaces = (cface.size() > maxfaces) ? cface.size() : maxfaces;
   }
 
-  Epetra_IntSerialDenseMatrix gids(ncell_use, maxfaces); // no Epetra_IntMultiVector :(
+  Epetra_IntSerialDenseMatrix gids(ncell_use,maxfaces); // no Epetra_IntMultiVector :(
 
   for (AmanziMesh::Entity_ID j = 0; j < ncell_own; ++j) {
-    mesh->cell_get_faces(j, &cface);
-    for (int k = 0; k < cface.size(); ++k) gids(j, k) = face_map.GID(cface[k]);
-    for (int k = cface.size(); k < maxfaces; ++k) gids(j, k) = 0;
+    mesh->getCellFaces(j, cface);
+    for (int k = 0; k < cface.size(); ++k)
+      gids(j,k) = face_map.GID(cface[k]);
+    for (int k = cface.size(); k < maxfaces; ++k)
+      gids(j,k) = 0;
   }
 
   // Import these GIDs to all used cells; sets values on ghost cells.
@@ -668,10 +665,10 @@ MeshLogicalAudit::check_cell_to_faces_ghost_data() const
 
   // Compare the ghost cell GIDs against the reference values just computed.
   for (AmanziMesh::Entity_ID j = ncell_own; j < ncell_use; ++j) {
-    mesh->cell_get_faces(j, &cface);
+    mesh->getCellFaces(j, cface);
     bool bad_data = false;
     for (int k = 0; k < cface.size(); ++k)
-      if (face_map.GID(cface[k]) != gids(j, k)) bad_data = true;
+      if (face_map.GID(cface[k]) != gids(j,k)) bad_data = true;
     if (bad_data) bad_cells.push_back(j);
   }
 
@@ -697,21 +694,20 @@ MeshLogicalAudit::check_cell_to_faces_ghost_data() const
 // that owns a particular face (node) must also own one of the cells containing
 // the face (node).
 
-bool
-MeshLogicalAudit::check_face_partition() const
+bool MeshLogicalAudit::check_face_partition() const
 {
   // Mark all the faces contained by owned cells.
   bool owned[nface];
   for (int j = 0; j < nface; ++j) owned[j] = false;
   AmanziMesh::Entity_ID_List cface;
-  for (AmanziMesh::Entity_ID j = 0; j < mesh->cell_map(false).NumMyElements(); ++j) {
-    mesh->cell_get_faces(j, &cface);
+  for (AmanziMesh::Entity_ID j = 0; j < mesh->getMap(AmanziMesh::Entity_kind::CELL, false).NumMyElements(); ++j) {
+    mesh->getCellFaces(j, cface);
     for (int k = 0; k < cface.size(); ++k) owned[cface[k]] = true;
   }
 
   // Verify that every owned face has been marked as belonging to an owned cell.
   AmanziMesh::Entity_ID_List bad_faces;
-  for (AmanziMesh::Entity_ID j = 0; j < mesh->face_map(false).NumMyElements(); ++j)
+  for (AmanziMesh::Entity_ID j = 0; j < mesh->getMap(AmanziMesh::Entity_kind::FACE, false).NumMyElements(); ++j)
     if (!owned[j]) bad_faces.push_back(j);
 
   if (!bad_faces.empty()) {
@@ -730,9 +726,7 @@ MeshLogicalAudit::check_face_partition() const
 // faces.  Implicitly assumes non-degenerate faces; the results are not
 // reliable otherwise.
 
-int
-MeshLogicalAudit::same_face(const AmanziMesh::Entity_ID_List fnode1,
-                            const AmanziMesh::Entity_ID_List fnode2) const
+int MeshLogicalAudit::same_face(const AmanziMesh::Entity_ID_List fnode1, const AmanziMesh::Entity_ID_List fnode2) const
 {
   int nn = fnode1.size();
 
@@ -749,45 +743,44 @@ MeshLogicalAudit::same_face(const AmanziMesh::Entity_ID_List fnode1,
 
   if (nn == 2) {
     // These are edges in a 2D mesh
-
+    
+    if (n == 0 && fnode1[1] == fnode2[1]) return 1;
+    if (n == 1 && fnode1[0] == fnode2[1]) return -1;
+    
     if (n == 0 && fnode1[1] == fnode2[1]) return 1;
     if (n == 1 && fnode1[0] == fnode2[1]) return -1;
 
-    if (n == 0 && fnode1[1] == fnode2[1]) return 1;
-    if (n == 1 && fnode1[0] == fnode2[1]) return -1;
-
-  } else {
+  }
+  else {
     for (i = 1; i < nn; ++i)
-      if (fnode1[(n + i) % nn] != fnode2[i]) break;
-    if (i == nn) return 1; // they match
+      if (fnode1[(n+i)%nn] != fnode2[i]) break;
+    if (i == nn) return 1;  // they match
 
     // Modify the permutation to reverse the orientation of fnode1.
-
+    
     for (i = 1; i < nn; ++i)
-      if (fnode1[(n - i + nn) % nn] != fnode2[i]) break;
-    if (i == nn) return -1; // matched nodes but orientation is reversed
+      if (fnode1[(n-i+nn)%nn] != fnode2[i]) break;
+    if (i == nn) return -1;   // matched nodes but orientation is reversed
   }
 
   return 0; // different faces
 }
 
 
-void
-MeshLogicalAudit::write_list(const AmanziMesh::Entity_ID_List& list, unsigned int max_out) const
+void MeshLogicalAudit::write_list(const AmanziMesh::Entity_ID_List &list, unsigned int max_out) const
 {
-  int num_out = min((unsigned int)list.size(), max_out);
+  int num_out = min((unsigned int) list.size(), max_out);
   for (int i = 0; i < num_out; ++i) os << " " << list[i];
-  if (num_out < list.size()) os << " [" << list.size() - num_out << " items omitted]";
+  if (num_out < list.size()) os << " [" << list.size()-num_out << " items omitted]";
   os << std::endl;
 }
 
-bool
-MeshLogicalAudit::global_any(bool value) const
+bool MeshLogicalAudit::global_any(bool value) const
 {
-  int lval = value, gval;
+  int lval=value, gval;
   comm_->MaxAll(&lval, &gval, 1);
   return gval;
 }
 
-} // namespace AmanziMesh
-} // namespace Amanzi
+} // close namespace Amanzi
+} // close namespace Amanzi
