@@ -14,12 +14,13 @@ namespace Amanzi {
 namespace AmanziMesh {
 
 class MeshLogicalAudit {
- public:
-  MeshLogicalAudit(const Teuchos::RCP<const AmanziMesh::Mesh>& mesh_, std::ostream& os = std::cout);
+public:
+
+  MeshLogicalAudit(const Teuchos::RCP<const AmanziMesh::Mesh> &mesh_, std::ostream& os=std::cout);
 
   // This is the main method.
   int Verify() const;
-
+  
   // The individual tests are also available.  While the tests are all formally
   // independent, there is an implicit order dependence of the tests in that a
   // test may assume certain mesh data has been verified, and that verification
@@ -38,9 +39,10 @@ class MeshLogicalAudit {
   bool check_cell_maps() const;
   bool check_cell_to_faces_ghost_data() const;
   bool check_face_partition() const;
-  bool check_cell_face_bisector_geometry() const;
+  bool check_cell_face_bisector_geometry() const;  
+  
+private:
 
- private:
   Teuchos::RCP<const AmanziMesh::Mesh> mesh;
 
   Comm_ptr_type comm_;
@@ -55,23 +57,18 @@ class MeshLogicalAudit {
   void write_list(const AmanziMesh::Entity_ID_List&, unsigned int) const;
   bool global_any(bool) const;
   int same_face(const AmanziMesh::Entity_ID_List, const AmanziMesh::Entity_ID_List) const;
-
+  
   bool check_maps(const Epetra_Map&, const Epetra_Map&) const;
   bool check_get_set_ids(AmanziMesh::Entity_kind) const;
   bool check_valid_set_id(AmanziMesh::Entity_kind) const;
   bool check_sets(AmanziMesh::Entity_kind, const Epetra_Map&, const Epetra_Map&) const;
-  bool check_get_set(AmanziMesh::Set_ID,
-                     AmanziMesh::Entity_kind,
-                     AmanziMesh::Parallel_type,
-                     const Epetra_Map&) const;
-  bool check_used_set(AmanziMesh::Set_ID,
-                      AmanziMesh::Entity_kind,
-                      const Epetra_Map&,
-                      const Epetra_Map&) const;
-
+  bool check_get_set(AmanziMesh::Set_ID, AmanziMesh::Entity_kind, AmanziMesh::Parallel_type, const Epetra_Map&) const;
+  bool check_used_set(AmanziMesh::Set_ID, AmanziMesh::Entity_kind, const Epetra_Map&, const Epetra_Map&) const;
+  
   // This is the vertex type for the test dependency graph.
-  typedef bool (MeshLogicalAudit::*Test)() const;
-  struct Vertex {
+  typedef bool (MeshLogicalAudit::* Test)() const;
+  struct Vertex
+  {
     Vertex() : run(true) {}
     std::string name;
     mutable bool run;
@@ -80,19 +77,17 @@ class MeshLogicalAudit {
 
   typedef boost::adjacency_list<boost::listS, boost::vecS, boost::directedS, Vertex> Graph;
   Graph g;
-
-  struct mark_do_not_run : public boost::bfs_visitor<> {
+  
+  struct mark_do_not_run : public boost::bfs_visitor<>
+  {
     template <class Vertex, class Graph>
-    void discover_vertex(Vertex v, Graph& gr)
-    {
-      gr[v].run = false;
-    }
+    void discover_vertex(Vertex v, Graph &gr) { gr[v].run = false; }
   };
-
+  
   void create_test_dependencies();
 };
 
-} // namespace AmanziMesh
-} // namespace Amanzi
+}
+} // close namespace Amanzi
 
 #endif
