@@ -15,64 +15,101 @@
 
 #pragma once
 
+#include "AmanziComm.hh"
+#include "GeometricModel.hh"
+#include "MeshFramework.hh"
+#include "MeshFrameworkFactory.hh"
 #include "Mesh.hh"
 #include "MeshFactory.hh"
-#include "GeometricModel.hh"
-#include "AmanziComm.hh"
+
 
 using namespace Amanzi;
 
-inline Teuchos::RCP<AmanziMesh::Mesh>
-createFrameworkStructuredUnitQuad(
-  const AmanziMesh::Preference& pref,
-  int nx,
-  int ny,
-  Comm_ptr_type comm = Teuchos::null,
-  const Teuchos::RCP<const AmanziGeometry::GeometricModel>& gm = Teuchos::null,
-  const Teuchos::RCP<Teuchos::ParameterList>& plist = Teuchos::null,
-  bool request_edges = false,
-  double dx = 1.0,
-  double dy = 1.0)
+inline
+Teuchos::RCP<AmanziMesh::MeshFramework> createFrameworkStructuredUnitQuad(
+  const AmanziMesh::Preference& pref, int nx, int ny,
+  Comm_ptr_type comm=Teuchos::null,
+  const Teuchos::RCP<const AmanziGeometry::GeometricModel>& gm=Teuchos::null,
+  const Teuchos::RCP<Teuchos::ParameterList>& plist=Teuchos::null,
+  double dx=1.0, double dy=1.0)
 {
   if (comm == Teuchos::null) comm = getDefaultComm();
-  AmanziMesh::MeshFactory fac(comm, gm, plist);
+  AmanziMesh::MeshFrameworkFactory fac(comm, gm, plist);
   fac.set_preference(pref);
-  return fac.create(0.0, 0.0, dx, dy, nx, ny, true, request_edges);
+  return fac.create(0.0,0.0, dx,dy, nx,ny);
+}
+
+inline
+Teuchos::RCP<AmanziMesh::Mesh>  createStructuredUnitQuad(
+  const AmanziMesh::Preference& pref, int nx, int ny,
+  Comm_ptr_type comm=Teuchos::null,
+  const Teuchos::RCP<const AmanziGeometry::GeometricModel>& gm=Teuchos::null,
+  const Teuchos::RCP<Teuchos::ParameterList>& plist=Teuchos::null,
+  double dx=1.0, double dy=1.0)
+{
+  auto mesh_fw = createFrameworkStructuredUnitQuad(pref, nx, ny, comm, gm, plist, dx, dy);
+  return Teuchos::rcp(new AmanziMesh::Mesh(mesh_fw));
 }
 
 
-inline Teuchos::RCP<AmanziMesh::Mesh>
-createFrameworkStructuredUnitHex(
+
+inline
+Teuchos::RCP<AmanziMesh::MeshFramework> createFrameworkStructuredUnitHex(
   const AmanziMesh::Preference& pref,
-  int nx,
-  int ny,
-  int nz,
-  Comm_ptr_type comm = Teuchos::null,
-  const Teuchos::RCP<const AmanziGeometry::GeometricModel>& gm = Teuchos::null,
-  const Teuchos::RCP<Teuchos::ParameterList>& plist = Teuchos::null,
-  bool request_edges = false,
-  double dx = 1.0,
-  double dy = 1.0,
-  double dz = 1.0)
+  int nx, int ny, int nz,
+  Comm_ptr_type comm=Teuchos::null,
+  const Teuchos::RCP<const AmanziGeometry::GeometricModel>& gm=Teuchos::null,
+  const Teuchos::RCP<Teuchos::ParameterList>& plist=Teuchos::null,
+  double dx=1.0, double dy=1.0, double dz=1.0)
 {
   if (comm == Teuchos::null) comm = getDefaultComm();
-  AmanziMesh::MeshFactory fac(comm, gm, plist);
+  AmanziMesh::MeshFrameworkFactory fac(comm, gm, plist);
   fac.set_preference(pref);
-  return fac.create(0.0, 0.0, 0.0, dx, dy, dz, nx, ny, nz, true, request_edges);
+  return fac.create(0.0,0.0,0.0,dx,dy,dz,nx,ny,nz);
+}
+
+inline
+Teuchos::RCP<AmanziMesh::Mesh> createStructuredUnitHex(
+  const AmanziMesh::Preference& pref,
+  int nx, int ny, int nz,
+  Comm_ptr_type comm=Teuchos::null,
+  const Teuchos::RCP<const AmanziGeometry::GeometricModel>& gm=Teuchos::null,
+  const Teuchos::RCP<Teuchos::ParameterList>& plist=Teuchos::null,
+  double dx=1.0, double dy=1.0, double dz=1.0)
+{
+  auto mesh_fw = createFrameworkStructuredUnitHex(pref, nx,ny,nz, comm, gm, plist, dx,dy,dz);
+  return Teuchos::rcp(new AmanziMesh::Mesh(mesh_fw));
 }
 
 
-inline Teuchos::RCP<AmanziMesh::Mesh>
-createFrameworkUnstructured(
+inline
+Teuchos::RCP<AmanziMesh::MeshFramework> createFrameworkUnstructured(
   const AmanziMesh::Preference& pref,
   const std::string& filename,
-  Comm_ptr_type comm = Teuchos::null,
-  const Teuchos::RCP<const AmanziGeometry::GeometricModel>& gm = Teuchos::null,
-  const Teuchos::RCP<Teuchos::ParameterList>& plist = Teuchos::null,
-  bool request_edges = false)
+  Comm_ptr_type comm=Teuchos::null,
+  const Teuchos::RCP<const AmanziGeometry::GeometricModel>& gm=Teuchos::null,
+  const Teuchos::RCP<Teuchos::ParameterList>& plist=Teuchos::null)
 {
   if (comm == Teuchos::null) comm = getDefaultComm();
+  AmanziMesh::MeshFrameworkFactory fac(comm, gm, plist);
+  fac.set_preference(pref);
+  return fac.create(filename);
+}
+
+inline
+Teuchos::RCP<AmanziMesh::Mesh> createUnstructured(
+  const AmanziMesh::Preference& pref,
+  const std::string& filename,
+  Comm_ptr_type comm=Teuchos::null,
+  const Teuchos::RCP<const AmanziGeometry::GeometricModel>& gm=Teuchos::null,
+  const Teuchos::RCP<Teuchos::ParameterList>& plist=Teuchos::null)
+{
   AmanziMesh::MeshFactory fac(comm, gm, plist);
   fac.set_preference(pref);
-  return fac.create(filename, true, request_edges);
+  return fac.create(filename);
+  // auto mesh_fw = createFrameworkUnstructured(pref, filename, comm, gm, plist);
+  // return Teuchos::rcp(new AmanziMesh::Mesh(mesh_fw));
 }
+
+
+
