@@ -19,6 +19,7 @@
 #include "AmanziComm.hh"
 #include "RegionFactory.hh"
 #include "Mesh.hh"
+#include "MeshFramework.hh"
 #include "MeshFactory.hh"
 #include "MeshAudit.hh"
 
@@ -39,35 +40,38 @@ TEST(MESH_FRACTURE_EXTRACTION_GENERATED)
   //
   // extract two inner surfaces as fractures
   std::vector<Framework> frameworks;
-  if (framework_enabled(Framework::MSTK)) { frameworks.push_back(Framework::MSTK); }
+  if (framework_enabled(Framework::MSTK)) {
+    frameworks.push_back(Framework::MSTK);
+  }
 
   for (const auto& frm : frameworks) {
     std::cout << std::endl
-              << "Testing Fracture Extraction with " << AmanziMesh::framework_names.at(frm)
-              << std::endl
+              << "Testing Fracture Extraction with " << AmanziMesh::to_string(frm) << std::endl
               << "------------------------------------------------" << std::endl;
-    auto parent_mesh = createFrameworkStructuredUnitHex(Preference{ frm }, 10, 10, 10, comm, gm);
+    auto parent_mesh = createFrameworkStructuredUnitHex(Preference{frm}, 10,10,10, comm, gm);
 
     // extract the fractures
-    std::vector<std::string> setnames{ "fracture 1", "fracture 2" };
+    std::vector<std::string> setnames{"fracture 1", "fracture 2"};
     MeshFactory fac(comm, gm);
-    fac.set_preference({ frm });
-    auto mesh = fac.create(parent_mesh, setnames, AmanziMesh::Entity_kind::FACE, false);
+    fac.set_preference({frm});
+    // Make cache of current mesh 
+    auto parent_mesh_cache = Teuchos::rcp(new Mesh(parent_mesh)); 
+    auto mesh = fac.create(parent_mesh_cache, setnames, AmanziMesh::Entity_kind::FACE, false);
 
     // test the surface mesh as a fracture mesh
     // -- mesh audit
     testMeshAudit<MeshAudit, Mesh>(mesh);
 
     // check we found the right number of things
-    int ncells =
-      mesh->num_entities(AmanziMesh::Entity_kind::CELL, AmanziMesh::Parallel_type::OWNED);
-    CHECK_CLOSE_SUMALL(10 * 10 * 2, ncells, *comm);
-    int nnodes =
-      mesh->num_entities(AmanziMesh::Entity_kind::NODE, AmanziMesh::Parallel_type::OWNED);
-    CHECK_CLOSE_SUMALL(11 * 11 * 2 - 11, nnodes, *comm);
-    int nfaces =
-      mesh->num_entities(AmanziMesh::Entity_kind::FACE, AmanziMesh::Parallel_type::OWNED);
-    CHECK_CLOSE_SUMALL(10 * 11 * 4 - 10, nfaces, *comm);
+    int ncells = mesh->getNumEntities(AmanziMesh::Entity_kind::CELL,
+            AmanziMesh::Parallel_type::OWNED);
+    CHECK_CLOSE_SUMALL(10*10*2, ncells , *comm);
+    int nnodes = mesh->getNumEntities(AmanziMesh::Entity_kind::NODE,
+            AmanziMesh::Parallel_type::OWNED);
+    CHECK_CLOSE_SUMALL(11*11*2-11, nnodes, *comm);
+    int nfaces = mesh->getNumEntities(AmanziMesh::Entity_kind::FACE,
+            AmanziMesh::Parallel_type::OWNED);
+    CHECK_CLOSE_SUMALL(10*11*4-10, nfaces, *comm);
   }
 }
 
@@ -90,47 +94,48 @@ TEST(MESH_FRACTURE_EXTRACTION_GENERATED_EXTRACTED_MANIFOLD)
   //
   // extract two inner surfaces as fractures
   std::vector<Framework> frameworks;
-  if (comm->NumProc() == 1) { frameworks.push_back(Framework::SIMPLE); }
-  if (framework_enabled(Framework::MSTK)) { frameworks.push_back(Framework::MSTK); }
+  if (comm->NumProc() == 1) {
+    frameworks.push_back(Framework::SIMPLE);
+  }
+  if (framework_enabled(Framework::MSTK)) {
+    frameworks.push_back(Framework::MSTK);
+  }
 
   for (const auto& frm : frameworks) {
     std::cout << std::endl
-              << "Testing Fracture Extraction with " << AmanziMesh::framework_names.at(frm)
-              << std::endl
+              << "Testing Fracture Extraction with " << AmanziMesh::to_string(frm) << std::endl
               << "------------------------------------------------" << std::endl;
-    auto parent_mesh = createFrameworkStructuredUnitHex(
-      Preference{ frm }, 10, 10, 10, comm, gm, Teuchos::null, true);
+    auto parent_mesh = createFrameworkStructuredUnitHex(Preference{frm}, 10,10,10,comm,gm,Teuchos::null,true);
 
     // extract the fractures
     auto fac_plist = Teuchos::rcp(new Teuchos::ParameterList());
 
     // Extracting multiple sets not supported by MeshExtractedManifold, see #596 part 4
     // std::vector<std::string> setnames{"fracture 1", "fracture 2"};
-    std::vector<std::string> setnames{ "fractures" };
-    fac_plist->sublist("unstructured")
-      .sublist("submesh")
-      .set<std::string>("extraction method", "manifold mesh");
+    std::vector<std::string> setnames{"fractures"};
+    fac_plist->sublist("unstructured").sublist("submesh").set<std::string>("extraction method", "manifold mesh");
     MeshFactory fac(comm, gm, fac_plist);
-    fac.set_preference({ frm });
-    auto mesh = fac.create(parent_mesh, setnames, AmanziMesh::Entity_kind::FACE, false);
+    fac.set_preference({frm});
+    // Make cache of current mesh 
+    auto parent_mesh_cache = Teuchos::rcp(new Mesh(parent_mesh)); 
+    auto mesh = fac.create(parent_mesh_cache, setnames, AmanziMesh::Entity_kind::FACE, false);
 
     // test the surface mesh as a fracture mesh
     // -- mesh audit
     testMeshAudit<MeshAudit, Mesh>(mesh);
 
     // check we found the right number of things
-    int ncells =
-      mesh->num_entities(AmanziMesh::Entity_kind::CELL, AmanziMesh::Parallel_type::OWNED);
-    CHECK_CLOSE_SUMALL(10 * 10 * 2, ncells, *comm);
-    int nnodes =
-      mesh->num_entities(AmanziMesh::Entity_kind::NODE, AmanziMesh::Parallel_type::OWNED);
-    CHECK_CLOSE_SUMALL(11 * 11 * 2 - 11, nnodes, *comm);
-    int nfaces =
-      mesh->num_entities(AmanziMesh::Entity_kind::FACE, AmanziMesh::Parallel_type::OWNED);
-    CHECK_CLOSE_SUMALL(10 * 11 * 4 - 10, nfaces, *comm);
+    int ncells = mesh->getNumEntities(AmanziMesh::Entity_kind::CELL,
+            AmanziMesh::Parallel_type::OWNED);
+    CHECK_CLOSE_SUMALL(10*10*2, ncells , *comm);
+    int nnodes = mesh->getNumEntities(AmanziMesh::Entity_kind::NODE,
+            AmanziMesh::Parallel_type::OWNED);
+    CHECK_CLOSE_SUMALL(11*11*2-11, nnodes, *comm);
+    int nfaces = mesh->getNumEntities(AmanziMesh::Entity_kind::FACE,
+            AmanziMesh::Parallel_type::OWNED);
+    CHECK_CLOSE_SUMALL(10*11*4-10, nfaces, *comm);
   }
 }
-
 
 TEST(MESH_FRACTURE_EXTRACTION_EXO)
 {
@@ -146,38 +151,38 @@ TEST(MESH_FRACTURE_EXTRACTION_EXO)
   //
   // extract two inner surfaces as fractures
   std::vector<Framework> frameworks;
-  if (framework_enabled(Framework::MSTK)) { frameworks.push_back(Framework::MSTK); }
+  if (framework_enabled(Framework::MSTK)) {
+    frameworks.push_back(Framework::MSTK);
+  }
 
   for (const auto& frm : frameworks) {
     std::cout << std::endl
-              << "Testing Fracture Extraction with " << AmanziMesh::framework_names.at(frm)
-              << std::endl
+              << "Testing Fracture Extraction with " << AmanziMesh::to_string(frm) << std::endl
               << "------------------------------------------------" << std::endl;
-    auto parent_mesh =
-      createFrameworkUnstructured(Preference{ frm }, "test/mesh_extracted_fracture.exo", comm, gm);
+    auto parent_mesh = createFrameworkUnstructured(Preference{frm}, "test/mesh_extracted_fracture.exo", comm, gm);
 
     // extract the fractures
     MeshFactory fac(comm, gm);
-    fac.set_preference({ frm });
-    auto mesh = fac.create(parent_mesh, { "fractures-two" }, AmanziMesh::Entity_kind::FACE, false);
+    fac.set_preference({frm});
+    auto parent_mesh_cache = Teuchos::rcp(new Mesh(parent_mesh));
+    auto mesh = fac.create(parent_mesh_cache, {"fractures-two"}, AmanziMesh::Entity_kind::FACE, false);
 
     // test the surface mesh as a fracture mesh
     // -- mesh audit
     testMeshAudit<MeshAudit, Mesh>(mesh);
 
     // check we found the right number of things
-    int ncells =
-      mesh->num_entities(AmanziMesh::Entity_kind::CELL, AmanziMesh::Parallel_type::OWNED);
-    CHECK_CLOSE_SUMALL(108, ncells, *comm);
-    int nnodes =
-      mesh->num_entities(AmanziMesh::Entity_kind::NODE, AmanziMesh::Parallel_type::OWNED);
+    int ncells = mesh->getNumEntities(AmanziMesh::Entity_kind::CELL,
+            AmanziMesh::Parallel_type::OWNED);
+    CHECK_CLOSE_SUMALL(108, ncells , *comm);
+    int nnodes = mesh->getNumEntities(AmanziMesh::Entity_kind::NODE,
+            AmanziMesh::Parallel_type::OWNED);
     CHECK_CLOSE_SUMALL(91, nnodes, *comm);
-    int nfaces =
-      mesh->num_entities(AmanziMesh::Entity_kind::FACE, AmanziMesh::Parallel_type::OWNED);
+    int nfaces = mesh->getNumEntities(AmanziMesh::Entity_kind::FACE,
+            AmanziMesh::Parallel_type::OWNED);
     CHECK_CLOSE_SUMALL(198, nfaces, *comm);
   }
 }
-
 
 TEST(MESH_FRACTURE_EXTRACTION_EXO_MANIFOLD)
 {
@@ -197,7 +202,9 @@ TEST(MESH_FRACTURE_EXTRACTION_EXO_MANIFOLD)
   //
   // extract two inner surfaces as fractures
   std::vector<Framework> frameworks;
-  if (framework_enabled(Framework::MSTK)) { frameworks.push_back(Framework::MSTK); }
+  if (framework_enabled(Framework::MSTK)) {
+    frameworks.push_back(Framework::MSTK);
+  }
 
   // NOTE, moab doesn't support edges, so this can't work here. See #596
   // if (framework_enabled(Framework::MOAB)) {
@@ -206,35 +213,36 @@ TEST(MESH_FRACTURE_EXTRACTION_EXO_MANIFOLD)
 
   for (const auto& frm : frameworks) {
     std::cout << std::endl
-              << "Testing Fracture Extraction with " << AmanziMesh::framework_names.at(frm)
-              << std::endl
+              << "Testing Fracture Extraction with " << AmanziMesh::to_string(frm) << std::endl
               << "------------------------------------------------" << std::endl;
-    auto parent_mesh = createFrameworkUnstructured(
-      Preference{ frm }, "test/mesh_extracted_fracture.exo", comm, gm, Teuchos::null, true);
+    auto parent_mesh = createFrameworkUnstructured(Preference{frm}, "test/mesh_extracted_fracture.exo", comm,gm,Teuchos::null);
 
     // extract the fractures
     auto fac_plist = Teuchos::rcp(new Teuchos::ParameterList());
-    std::vector<std::string> setnames{ "fractures" };
-    fac_plist->sublist("unstructured")
-      .sublist("submesh")
-      .set<std::string>("extraction method", "manifold mesh");
+    std::vector<std::string> setnames{"fractures"};
+    fac_plist->sublist("unstructured").sublist("submesh").set<std::string>("extraction method", "manifold mesh");
     MeshFactory fac(comm, gm, fac_plist);
-    fac.set_preference({ frm });
-    auto mesh = fac.create(parent_mesh, setnames, AmanziMesh::Entity_kind::FACE, false);
+    fac.set_preference({frm});
+    auto parent_mesh_cache = Teuchos::rcp(new Mesh(parent_mesh)); 
+    auto mesh = fac.create(parent_mesh_cache, setnames, AmanziMesh::Entity_kind::FACE, false);
 
     // test the surface mesh as a fracture mesh
     // -- mesh audit
     testMeshAudit<MeshAudit, Mesh>(mesh);
 
     // check we found the right number of things
-    int ncells =
-      mesh->num_entities(AmanziMesh::Entity_kind::CELL, AmanziMesh::Parallel_type::OWNED);
-    CHECK_CLOSE_SUMALL(108, ncells, *comm);
-    int nnodes =
-      mesh->num_entities(AmanziMesh::Entity_kind::NODE, AmanziMesh::Parallel_type::OWNED);
+    int ncells = mesh->getNumEntities(AmanziMesh::Entity_kind::CELL,
+            AmanziMesh::Parallel_type::OWNED);
+    CHECK_CLOSE_SUMALL(108, ncells , *comm);
+    int nnodes = mesh->getNumEntities(AmanziMesh::Entity_kind::NODE,
+            AmanziMesh::Parallel_type::OWNED);
     CHECK_CLOSE_SUMALL(91, nnodes, *comm);
-    int nfaces =
-      mesh->num_entities(AmanziMesh::Entity_kind::FACE, AmanziMesh::Parallel_type::OWNED);
+    int nfaces = mesh->getNumEntities(AmanziMesh::Entity_kind::FACE,
+            AmanziMesh::Parallel_type::OWNED);
     CHECK_CLOSE_SUMALL(198, nfaces, *comm);
   }
 }
+
+
+
+
