@@ -13,27 +13,35 @@
 
 #include "AmanziComm.hh"
 #include "Mesh.hh"
-#include "MeshFactory.hh"
 #include "MeshAudit.hh"
 
 #include "framework_meshes.hh"
 #include "geometry_harnesses.hh"
 
+//
+// This tests the geometric routines in MeshCache, without caching, and with an
+// existing framework.
+//
+// Effectively this means that all function calls that need a framework will
+// use it, and the rest will call the algorithm.
+//
 TEST(MESH_GEOMETRY_PLANAR)
 {
   // a 2D, generated, structured quad on the unit square, NX=NY=2
   std::vector<Framework> frameworks;
   // works in MSTK
-  if (framework_enabled(Framework::MSTK)) { frameworks.push_back(Framework::MSTK); }
+  if (framework_enabled(Framework::MSTK)) {
+    frameworks.push_back(Framework::MSTK);
+  }
 
   for (const auto& frm : frameworks) {
     std::cout << std::endl
-              << "Testing 2D geometry with " << AmanziMesh::framework_names.at(frm) << std::endl
+              << "Testing 2D geometry with " << AmanziMesh::to_string(frm) << std::endl
               << "------------------------------------------------" << std::endl;
 
-    auto mesh = createFrameworkStructuredUnitQuad(Preference{ frm }, 2, 2);
+    auto mesh = createStructuredUnitQuad(Preference{frm}, 2,2);
     testMeshAudit<MeshAudit, Mesh>(mesh);
-    testGeometryQuad(mesh, 2, 2);
+    testGeometryQuad(mesh,2,2);
     testExteriorMapsUnitBox(mesh, 2, 2);
   }
 }
@@ -47,19 +55,22 @@ TEST(MESH_GEOMETRY_1CUBE_GENERATED)
 
   std::vector<Framework> frameworks;
   // works in MSTK & SIMPLE (in serial)
-  if (framework_enabled(Framework::MSTK)) { frameworks.push_back(Framework::MSTK); }
-  if (getDefaultComm()->NumProc() == 1) frameworks.push_back(Framework::SIMPLE);
+  if (framework_enabled(Framework::MSTK)) {
+    frameworks.push_back(Framework::MSTK);
+  }
+  if (getDefaultComm()->NumProc() == 1)
+    frameworks.push_back(Framework::SIMPLE);
 
   for (const auto& frm : frameworks) {
     std::cout << std::endl
-              << "Testing 3D geometry with " << AmanziMesh::framework_names.at(frm) << std::endl
+              << "Testing 3D geometry with " << AmanziMesh::to_string(frm) << std::endl
               << "------------------------------------------------" << std::endl;
-    auto mesh = createFrameworkStructuredUnitHex(Preference{ frm }, 1, 1, 1);
+    auto mesh = createStructuredUnitHex(Preference{frm}, 1,1,1);
     testMeshAudit<MeshAudit, Mesh>(mesh);
-    testGeometryCube<Mesh>(mesh, 1, 1, 1);
+    testGeometryCube(mesh,1,1,1);
 
     // Exterior maps not supported by SIMPLE
-    if (frm != Framework::SIMPLE) testExteriorMapsUnitBox(mesh, 1, 1, 1);
+    if (frm != Framework::SIMPLE) testExteriorMapsUnitBox(mesh, 1,1,1);
   }
 }
 
@@ -72,18 +83,23 @@ TEST(MESH_GEOMETRY_1CUBE_EXO)
 
   // works in MSTK or MOAB
   std::vector<Framework> frameworks;
-  if (framework_enabled(Framework::MSTK)) { frameworks.push_back(Framework::MSTK); }
-  if (framework_enabled(Framework::MOAB)) { frameworks.push_back(Framework::MOAB); }
+  if (framework_enabled(Framework::MSTK)) {
+    frameworks.push_back(Framework::MSTK);
+  }
+  if (framework_enabled(Framework::MOAB)) {
+    frameworks.push_back(Framework::MOAB);
+  }
 
   for (const auto& frm : frameworks) {
     std::cout << std::endl
-              << "Testing 3D Box 1x1x1 Exo geometry with " << AmanziMesh::framework_names.at(frm)
-              << std::endl
+              << "Testing 3D Box 1x1x1 Exo geometry with " << AmanziMesh::to_string(frm) << std::endl
               << "------------------------------------------------" << std::endl;
-    auto mesh = createFrameworkUnstructured(Preference{ frm }, "test/hex_1x1x1_sets.exo");
+    auto mesh = createUnstructured(Preference{frm}, "test/hex_1x1x1_sets.exo");
     testMeshAudit<MeshAudit, Mesh>(mesh);
-    testGeometryCube<Mesh>(mesh, 1, 1, 1);
-    if (frm == Framework::MSTK) testExteriorMapsUnitBox(mesh, 1, 1, 1);
+    testGeometryCube(mesh,1,1,1);
+    if (frm == Framework::MSTK)
+      testExteriorMapsUnitBox(mesh, 1,1,1);
+
   }
 }
 
@@ -93,17 +109,22 @@ TEST(MESH_GEOMETRY_3CUBE)
   // a 3D, generated, structured hex on the unit cube, NX=NY=NZ=3
   // works in MSTK & SIMPLE (in serial)
   std::vector<Framework> frameworks;
-  if (framework_enabled(Framework::MSTK)) { frameworks.push_back(Framework::MSTK); }
-  if (getDefaultComm()->NumProc() == 1) frameworks.push_back(Framework::SIMPLE);
+  if (framework_enabled(Framework::MSTK)) {
+    frameworks.push_back(Framework::MSTK);
+  }
+  if (getDefaultComm()->NumProc() == 1)
+    frameworks.push_back(Framework::SIMPLE);
 
   for (const auto& frm : frameworks) {
     std::cout << std::endl
-              << "Testing 3D Box 3x3x3 with " << AmanziMesh::framework_names.at(frm) << std::endl
+              << "Testing 3D Box 3x3x3 with " << AmanziMesh::to_string(frm) << std::endl
               << "------------------------------------------------" << std::endl;
-    auto mesh = createFrameworkStructuredUnitHex(Preference{ frm }, 3, 3, 3);
+    auto mesh = createStructuredUnitHex(Preference{frm}, 3,3,3);
     testMeshAudit<MeshAudit, Mesh>(mesh);
-    testGeometryCube<Mesh>(mesh, 3, 3, 3);
-    if (frm == Framework::MSTK) { testExteriorMapsUnitBox(mesh, 3, 3, 3); }
+    testGeometryCube(mesh,3,3,3);
+    if (frm == Framework::MSTK) {
+      testExteriorMapsUnitBox(mesh,3,3,3);
+    }
   }
 }
 
@@ -113,21 +134,25 @@ TEST(MESH_GEOMETRY_3CUBE_EXO)
   // a 3D exodus file, structured hex on the unit cube, NX=NY=NZ=3
   // works in MSTK or MOAB
   std::vector<Framework> frameworks;
-  if (framework_enabled(Framework::MSTK)) { frameworks.push_back(Framework::MSTK); }
-  if (framework_enabled(Framework::MOAB) && getDefaultComm()->NumProc() == 1) {
+  if (framework_enabled(Framework::MSTK)) {
+    frameworks.push_back(Framework::MSTK);
+  }
+  if (framework_enabled(Framework::MOAB) &&
+      getDefaultComm()->NumProc() == 1) {
     // moab only reads exo in serial, otherwise must read par
     frameworks.push_back(Framework::MOAB);
   }
 
   for (const auto& frm : frameworks) {
     std::cout << std::endl
-              << "Testing 3D Box 3x3x3 Exo with " << AmanziMesh::framework_names.at(frm)
-              << std::endl
+              << "Testing 3D Box 3x3x3 Exo with " << AmanziMesh::to_string(frm) << std::endl
               << "------------------------------------------------" << std::endl;
-    auto mesh = createFrameworkUnstructured(Preference{ frm }, "test/hex_3x3x3_sets.exo");
+    auto mesh = createUnstructured(Preference{frm}, "test/hex_3x3x3_sets.exo");
     testMeshAudit<MeshAudit, Mesh>(mesh);
-    testGeometryCube<Mesh>(mesh, 3, 3, 3);
-    if (frm == Framework::MSTK) { testExteriorMapsUnitBox(mesh, 3, 3, 3); }
+    testGeometryCube(mesh,3,3,3);
+    if (frm == Framework::MSTK) {
+      testExteriorMapsUnitBox(mesh,3,3,3);
+    }
   }
 }
 
@@ -152,11 +177,11 @@ TEST(MESH_GEOMETRY_3CUBE_EXO)
 
 //   for (const auto& frm : frameworks) {
 //     std::cout << std::endl
-//               << "Testing 3D Box 3x3x3 Exo with " << AmanziMesh::framework_names.at(frm) << std::endl
+//               << "Testing 3D Box 3x3x3 Exo with " << AmanziMesh::to_string(frm) << std::endl
 //               << "------------------------------------------------" << std::endl;
-//     auto mesh = createFrameworkUnstructured(Preference{frm}, "test/hex_3x3x3.par");
+//     auto mesh = createUnstructured(Preference{frm}, "test/hex_3x3x3.par");
 //     testMeshAudit<MeshAudit, Mesh>(mesh);
-//     testGeometryCube<Mesh>(mesh,3,3,3);
+//     testGeometryCube(mesh,3,3,3);
 //     if (frm == Framework::MSTK) {
 //       testExteriorMapsUnitBox(mesh,3,3,3);
 //     }
@@ -169,18 +194,22 @@ TEST(MESH_GEOMETRY_2x3CUBE)
   // a 3D, generated, structured hex on the unit cube, NX=NY=NZ=3
   // works in MSTK & SIMPLE (in serial)
   std::vector<Framework> frameworks;
-  if (framework_enabled(Framework::MSTK)) { frameworks.push_back(Framework::MSTK); }
-  if (getDefaultComm()->NumProc() == 1) frameworks.push_back(Framework::SIMPLE);
+  if (framework_enabled(Framework::MSTK)) {
+    frameworks.push_back(Framework::MSTK);
+  }
+  if (getDefaultComm()->NumProc() == 1)
+    frameworks.push_back(Framework::SIMPLE);
 
   for (const auto& frm : frameworks) {
     std::cout << std::endl
-              << "Testing 3D Box 2x2x3 with " << AmanziMesh::framework_names.at(frm) << std::endl
+              << "Testing 3D Box 2x2x3 with " << AmanziMesh::to_string(frm) << std::endl
               << "------------------------------------------------" << std::endl;
-    auto mesh = createFrameworkStructuredUnitHex(Preference{ frm }, 2, 2, 3);
+    auto mesh = createStructuredUnitHex(Preference{frm}, 2,2,3);
     testMeshAudit<MeshAudit, Mesh>(mesh);
-    testGeometryCube<Mesh>(mesh, 2, 2, 3);
+    testGeometryCube(mesh,2,2,3);
 
-    if (frm == Framework::MSTK) testExteriorMapsUnitBox(mesh, 2, 2, 3);
+    if (frm == Framework::MSTK)
+      testExteriorMapsUnitBox(mesh,2,2,3);
   }
 }
 
@@ -191,7 +220,9 @@ TEST(MESH_GEOMETRY_FRACTURE_EXO)
   // Note this only checks the exo mesh, which does not have the fractures in it!
   // Actual fracture capability is checked later, but would depend on this test passing.
   std::vector<Framework> frameworks;
-  if (framework_enabled(Framework::MSTK)) { frameworks.push_back(Framework::MSTK); }
+  if (framework_enabled(Framework::MSTK)) {
+    frameworks.push_back(Framework::MSTK);
+  }
 
   // Not sure what is up with this mesh, but MOAB errors trying to read it.
   // if (framework_enabled(Framework::MOAB)) {
@@ -200,9 +231,9 @@ TEST(MESH_GEOMETRY_FRACTURE_EXO)
 
   for (const auto& frm : frameworks) {
     std::cout << std::endl
-              << "Testing 3D Fracture Exo with " << AmanziMesh::framework_names.at(frm) << std::endl
+              << "Testing 3D Fracture Exo with " << AmanziMesh::to_string(frm) << std::endl
               << "------------------------------------------------" << std::endl;
-    auto mesh = createFrameworkUnstructured(Preference{ frm }, "test/fractures.exo");
+    auto mesh = createUnstructured(Preference{frm}, "test/fractures.exo");
     testMeshAudit<MeshAudit, Mesh>(mesh);
   }
 }
@@ -212,13 +243,16 @@ TEST(MESH_GEOMETRY_PINCHOUTS)
 {
   // only MSTK can handle this mesh -- it has degeneracies
   std::vector<Framework> frameworks;
-  if (framework_enabled(Framework::MSTK)) { frameworks.push_back(Framework::MSTK); }
+  if (framework_enabled(Framework::MSTK)) {
+    frameworks.push_back(Framework::MSTK);
+  }
 
   for (const auto& frm : frameworks) {
     std::cout << std::endl
-              << "Testing 3D Pinchout with " << AmanziMesh::framework_names.at(frm) << std::endl
+              << "Testing 3D Pinchout with " << AmanziMesh::to_string(frm) << std::endl
               << "------------------------------------------------" << std::endl;
-    auto mesh = createFrameworkUnstructured(Preference{ frm }, "test/test_pri_pinchout_mesh.exo");
+    auto mesh = createUnstructured(Preference{frm}, "test/test_pri_pinchout_mesh.exo");
     testMeshAudit<MeshAudit, Mesh>(mesh);
   }
 }
+
